### 2025-07-14 #8583 @editorialbot
[PRE REVIEW]: h-NUMO: A Hydrostatic Version of the Non-hydrostatic Model of the Ocean
**Submitting author:** @ygahounzo (<a href="http://orcid.org/0000-0001-8178-4887">Yao Gahounzo</a>)
**Repository:** https://github.com/ygahounzo/h-NUMO
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
