### 2025-07-16 07fa25a
Update on the paper
