+++ paper/paper.bib
@@ -13,72 +13,116 @@
 }
 
 @article{abdi2016efficient,
-  title     = {Efficient construction of unified continuous and discontinuous {Galerkin} formulations for the 3D {E}uler equations},
-  author    = {Abdi, Daniel S and Giraldo, Francis X},
-  journal   = {Journal of Computational Physics},
-  volume    = {320},
-  pages     = {46--68},
-  year      = {2016},
-  publisher = {Elsevier}
+  title = {Efficient construction of unified continuous and discontinuous Galerkin formulations for the 3D Euler equations},
+  journal = {Journal of Computational Physics},
+  volume = {320},
+  pages = {46-68},
+  year = {2016},
+  issn = {0021-9991},
+  doi = {https://doi.org/10.1016/j.jcp.2016.05.033},
+  url = {https://www.sciencedirect.com/science/article/pii/S0021999116301784},
+  author = {Daniel S. Abdi and Francis X. Giraldo},
+  keywords = {NUMA, Continuous Galerkin, Discontinuous Galerkin},
+  abstract = {A unified approach for the numerical solution of the 3D hyperbolic Euler equations using high order methods, namely continuous Galerkin (CG) and discontinuous Galerkin (DG) methods, is presented. First, we examine how classical CG that uses a global storage scheme can be constructed within the DG framework using constraint imposition techniques commonly used in the finite element literature. Then, we implement and test a simplified version in the Non-hydrostatic Unified Model of the Atmosphere (NUMA) for the case of explicit time integration and a diagonal mass matrix. Constructing CG within the DG framework allows CG to benefit from the desirable properties of DG such as, easier hp-refinement, better stability etc. Moreover, this representation allows for regional mixing of CG and DG depending on the flow regime in an area. The different flavors of CG and DG in the unified implementation are then tested for accuracy and performance using a suite of benchmark problems representative of cloud-resolving scale, meso-scale and global-scale atmospheric dynamics. The value of our unified approach is that we are able to show how to carry both CG and DG methods within the same code and also offer a simple recipe for modifying an existing CG code to DG and vice versa.}
 }
 
 @article{kopera2023non,
   title     = {Non-hydrostatic unified model of the ocean with application to ice/ocean interaction modeling},
   author    = {Kopera, Michal A and Gahounzo, Yao and Enderlin, Ellyn M and Giraldo, Francis X and Maslowski, Wieslaw},
   journal   = {GEM-International Journal on Geomathematics},
+  url = {https://doi.org/10.1007/s13137-022-00212-7},
+  doi = {10.1007/s13137-022-00212-7},
   volume    = {14},
   number    = {1},
   pages     = {2},
   year      = {2023},
-  publisher = {Springer}
+  publisher = {Springer},
+  abstract = {The non-hydrostatic unified model of the ocean (NUMO) has been developed to advance model capability to realistically represent the dynamics and ice/ocean interactions in Greenland fjords, including an accurate representation of complex fjord geometries. To that end, NUMO uses high-order spectral element methods on unstructured grids to solve the incompressible Navier–Stokes equations complemented with heat and salinity transport equations. This paper presents the model’s description and discusses the formulation of ice/ocean Neumann boundary conditions based on the three-equation model. We validate the model on a range of test cases. The convergence study on the classical Kovasznay flow shows exponential convergence with arbitrary basis function polynomial order. The lock-exchange and density current cases show that the model results of buoyancy-driven flows solved with 2D and 3D unstructured meshes agree well with previously published findings. Finally, we show that a high-order simulation of an ice block immersed in saline water produces results that match both direct numerical simulation and laboratory experiments.}
 }
 
 @article{escobar2012high,
[…]
