### 2025-09-30 #8583 @mancellin
Sorry for the late answer.
Yes, I'm interested in reviewing this paper.
