### 2025-10-06 #9147 @editorialbot
[REVIEW]: h-NUMO: A high-performance discontinuous-Galerkin code for hydrostatic ocean modelling
**Submitting author:** @ygahounzo (<a href="http://orcid.org/0000-0001-8178-4887">Yao Gahounzo</a>)
**Repository:** https://github.com/ygahounzo/h-NUMO
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
