### 2025-11-02 f43661f
Add option to compile the code in Debian/Ubuntu, see the documentation
