### 2025-11-08 #9147 @ygahounzo
@ZhiLiHydro, this work started about 3 years ago. I have submitted a longer paper on this work to the Journal of Computational Physics, describing all the mathematics and numerical results, and I conducted a comparative analysis with the Hybrid Coordinate Ocean Model (HYCOM), which was published about a week ago. In that paper, I referenced the h-NUMO GitHub repo; however, the publication was not explicitly about the code.

Also, I recently updated the documentation, adding instructions for compiling the code on Mac and Debian/Ubuntu.
