### 2025-11-12 cc50ffa
Removed some unnecessary initialization
