+++ src/initial_conditions.F90
@@ -4,13 +4,11 @@
 !>@date March 27, 2023
 !-----------------------------------------------------------------!
 
-subroutine initial_conditions(q_df, pbprime_init, pbprime_df, pbprime_face, &
-   one_over_pbprime, one_over_pbprime_face, pbprime_edge, one_over_pbprime_edge, &
-   one_over_pbprime_df, qb_df, qprime_df, alpha, one_over_pbprime_df_face, pbprime_df_face, &
-   zbot_df, tau_wind_df, z_interface)
+subroutine initial_conditions(q_df, pbprime_df, qb_df, alpha, pbprime_df_face, zbot_df, &
+   tau_wind_df, z_interface)
     
 
-   use mod_grid, only: nelem, nface, npoin_q, npoin, coord
+   use mod_grid, only: nelem, nface, npoin_q, npoin, coord, intma_dg_quad, face
    use mod_constants, only: gravity, pi, tol, omega, earth_radius
    use mod_initial_mlswe, only: interpolate_from_dof_to_quad_uv_init, interpolate_pbprime_init
    use mod_basis, only: ngl, nq
@@ -21,21 +19,14 @@ subroutine initial_conditions(q_df, pbprime_init, pbprime_df, pbprime_face, &
    use mod_types, only : r8
    use mpi
    use mod_mpi_utilities, only: MPI_PRECISION
+   use mod_face, only: imapl_q
 
    implicit none
    
    real, dimension(3,npoin, nlayers) :: q_df
-   real, dimension(3,npoin, nlayers) :: qprime_df
-   real, dimension(npoin_q) :: pbprime_init
    real, dimension(npoin) :: pbprime_df
-   real, dimension(2,nq,nface) :: pbprime_face
-   real, dimension(npoin_q) :: one_over_pbprime
-   real, dimension(2,nq,nface) :: one_over_pbprime_face
-   real, dimension(2,ngl,nface) :: one_over_pbprime_df_face
    real, dimension(2,ngl,nface) :: pbprime_df_face
-   real, dimension(nq, nface) :: pbprime_edge
-   real, dimension(nq, nface) :: one_over_pbprime_edge
-   real, dimension(npoin) :: one_over_pbprime_df, zbot_df
+   real, dimension(npoin) :: zbot_df
    real, dimension(4,npoin) :: qb_df
    real, dimension(nlayers) :: alpha
    real, dimension(2,npoin) :: tau_wind_df
@@ -48,7 +39,7 @@ subroutine initial_conditions(q_df, pbprime_init, pbprime_df, pbprime_face, &
 
    integer :: e, m, n, I1, iquad, Iq
    real :: x, y, xmid, ymid, L, amp, r, rho_0, H_bot
-   integer :: ierr, iface, ilr, k,ip
+   integer :: ierr, iface, ilr, k,ip, il, jl, kl, el
    real :: xmax, xmin, ymax, ymin, zmin, zmax, yl, xm
    real :: xmax_l, xmin_l, ymax_l, ymin_l, zmin_l, zmax_l, Ly
    real :: delta
@@ -68,15 +59,9 @@ subroutine initial_conditions(q_df, pbprime_init, pbprime_df, pbprime_face, &
    z_interface = 0.0
    pbprime_df = 0.0
    q_df = 0.0
-   one_over_pbprime = 0.0
-   one_over_pbprime_face = 0.0
-   pbprime_edge = 0.0
-   one_over_pbprime_edge = 0.0
-   one_over_pbprime_df = 0.0
    u_df = 0.0
    v_df = 0.0
    tau_wind_df = 0.0
-   one_over_pbprime_df_face = 0.0
    pbprime_df_face = 0.0
 
[…]
