+++ src/mod_rhs_btp.F90
@@ -20,8 +20,7 @@ module mod_rhs_btp
                            zetaq_x, zetaq_y, zetaq_z, &
                            jacq, massinv
 
-    public :: create_rhs_btp, create_rhs_btp_momentum, &
-              create_rhs_btp_mom_mass, btp_mass_advection_rhs
+    public :: create_rhs_btp
 
 contains
 
@@ -58,54 +57,13 @@ contains
 
     end subroutine create_rhs_btp
 
-    subroutine create_rhs_btp_momentum(rhs_mom,qb,qb_face)
-
-        implicit none
-
-        real, dimension(2, npoin), intent(out) :: rhs_mom
-        real, dimension(4,npoin_q), intent(in) :: qb
-        real, dimension(4, 2, nq, nface), intent(in) :: qb_face
-        
-        integer :: I
-
-        call create_rhs_btp_dynamics_volume(rhs_mom,qb)
-
-        call Apply_btp_fluxes(rhs_mom,qb_face)
-
-    end subroutine create_rhs_btp_momentum
-
-    subroutine create_rhs_btp_mom_mass(rhs,qb,qb_face)
-
-        implicit none
-
-        real, dimension(3, npoin), intent(out) :: rhs
-        real, dimension(4,npoin_q), intent(in) :: qb
-        real, dimension(4, 2, nq, nface), intent(in) :: qb_face
-
-        call create_rhs_btp_dynamics_volume_all(rhs, qb)
-
-        call Apply_btp_fluxes_all(rhs,qb_face)
-
-    end subroutine create_rhs_btp_mom_mass
-
-    subroutine btp_mass_advection_rhs(pb_advec)
-
-        implicit none
-
-        real, dimension(npoin), intent(out) :: pb_advec
-
-        call create_rhs_btp_dynamics_volume_mass(pb_advec)
-        call  Apply_btp_flux_mass(pb_advec)
-
-    end subroutine btp_mass_advection_rhs
-
     subroutine create_rhs_btp_volume_qdf(rhs, qb_df, qprime_df)
 
         use mod_grid, only : npoin_q, npoin, nelem, intma_dg_quad, intma
         use mod_basis, only: npts
         use mod_constants, only: gravity
         use mod_initial, only: grad_zbot_quad, tau_wind, psih, dpsidx,dpsidy, indexq, wjac, &
-                                coriolis_quad, one_over_pbprime, alpha_mlswe
+                                pbprime_df, coriolis_quad, alpha_mlswe
         use mod_variables, only: tau_bot_ave, H_ave, Qu_ave, Quv_ave, Qv_ave, ope_ave, &
                                 uvb_ave, btp_mass_flux_ave, ope2_ave
         use mod_variables, only: Q_uu_dp, Q_uv_dp, Q_vv_dp, H_bcl
@@ -120,7 +78,7 @@ contains
         real :: sc_x, sc_y, Hq, qu, quv, qv
         real :: wq, hi, dhdx, dhdy, coef_fric, tb_u, tb_v, ope, &
                 dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd
-        real :: pp, up, vp
+        real :: pp, up, vp, pbq
 
         integer :: I, Iq, ip
 
@@ -136,7 +94,7 @@ contains
         do Iq = 1, npoin_q
 
             dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0
-            pp = 0.0; up = 0.0; vp = 0.0
+            pp = 0.0; up = 0.0; vp = 0.0; pbq = 0.0
 
             do ip = 1,npts
                 I = indexq(ip,Iq)
@@ -149,6 +107,7 @@ contains
[…]
