+++ output/run_script_serial.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cp ../bin/numo3d .
+
+./numo3d

[…]
