### 2025-11-28 1610acc
Update on barotropic communication routines
