+++ src/mod_rhs_btp.F90
@@ -24,7 +24,7 @@ module mod_rhs_btp
 
 contains
 
-    subroutine create_rhs_btp(rhs,qb_df,qprime_df)
+    subroutine create_rhs_btp_v0(rhs,qb_df,qprime_df)
 
         implicit none
 
@@ -42,7 +42,40 @@ contains
 
         call btp_create_postcommunicator(qb_df_face,4)
 
-        call creat_btp_fluxes_qdf(rhs,qb_df_face)
+        call create_btp_fluxes_qdf(rhs,qb_df_face)
+
+        ! Compute RHS viscosity terms
+
+        if(method_visc == 1) then
+            call btp_create_laplacian_v2(rhs_visc_btp, qprime_df, qb_df)
+        else
+            call btp_create_laplacian(rhs_visc_btp,qb_df)
+        endif 
+
+        rhs(2,:) = rhs(2,:) + rhs_visc_btp(1,:)
+        rhs(3,:) = rhs(3,:) + rhs_visc_btp(2,:)
+
+    end subroutine create_rhs_btp_v0
+
+    subroutine create_rhs_btp(rhs,qb_df,qprime_df)
+
+        implicit none
+
+        real, dimension(3, npoin), intent(out) :: rhs
+        real, dimension(4,npoin), intent(in) :: qb_df
+        real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
+        real, dimension(4, 2, ngl, nface) :: qb_df_face
+        real, dimension(2,npoin) :: rhs_visc_btp
+
+        ! call btp_extract_df(qb_df_face, qb_df)
+
+        call btp_create_precommunicator_v1(qb_df,4)
+
+        call create_rhs_btp_volume_qdf(rhs, qb_df, qprime_df)
+
+        call create_btp_fluxes_qdf_v1(rhs,qb_df)
+
+        call btp_create_postcommunicator_v1(rhs,4)
 
         ! Compute RHS viscosity terms
 
@@ -167,7 +200,7 @@ contains
 
     end subroutine create_rhs_btp_volume_qdf
 
-    subroutine creat_btp_fluxes_qdf(rhs,qb_df_face)
+    subroutine create_btp_fluxes_qdf(rhs,qb_df_face)
 
         use mod_basis, only: psiq, ngl
         use mod_grid, only:  npoin, intma, nface,face
@@ -332,6 +365,188 @@ contains
         rhs(2,:) = massinv(:)*rhs(2,:)
         rhs(3,:) = massinv(:)*rhs(3,:)
 
-    end subroutine creat_btp_fluxes_qdf
+    end subroutine create_btp_fluxes_qdf
+
+    subroutine create_btp_fluxes_qdf_v1(rhs,qb)
+
+        use mod_basis, only: psiq, ngl
+        use mod_grid, only:  npoin, intma, nface,face
+        use mod_face, only: imapl, imapr, normal_vector_q, jac_faceq
+        use mod_variables, only: H_face_ave,ope_face_ave,btp_mass_flux_face_ave, &
+                                Qu_face_ave, Qv_face_ave, one_plus_eta_edge_2_ave, &
+                                Q_uu_dp_edge, Q_uv_dp_edge, Q_vv_dp_edge, &
+                                uvb_face_ave, H_bcl_edge, ope2_face_ave
+        use mod_initial, only: alpha_mlswe, pbprime_df_face, pbprime_df
+        use mod_constants, only: gravity
+        use mod_input, only: nlayers
+
+        implicit none
+
+        real, dimension(3, npoin), intent(inout)  :: rhs
+        real, dimension(4, npoin), intent(in) :: qb
+
[…]
