+++ src/create_rhs_dynamics_flux.F90
@@ -181,6 +181,215 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
  
  end subroutine create_nbhs_face_df
 
+  subroutine create_nbhs_face_df_v1(rhs,q_send,q_recv,nvarb,multirate)
+
+   use mod_basis, only: ngl, FACE_CHILDREN,nq, psiq
+ 
+   use mod_face, only: normal_vector_q, jac_faceq, imapl, imapr, face_send
+ 
+   use mod_grid, only: nelem, npoin, intma, face, nboun, mod_grid_get_face_nq, face_type,nface
+ 
+   use mod_initial, only: nvar, alpha_mlswe
+ 
+   use mod_metrics, only: jac, massinv
+ 
+   use mod_p4est, only: scatter_element_2d, scatter_element_2d_subface, &
+                        gather_element_2d_subface, plist, lev_list
+ 
+   use mod_parallel, only: nbh_send_recv, nbh_send_recv_multi, nbh_send_recv_half, &
+                           num_nbh, num_send_recv
+ 
+   use mod_ref, only: nmessage
+
+   use mod_input, only: nlayers
+   use mod_constants, only: gravity
+   use mod_variables, only: H_face_ave,ope_face_ave,btp_mass_flux_face_ave, &
+                                Qu_face_ave, Qv_face_ave, one_plus_eta_edge_2_ave, &
+                                Q_uu_dp_edge, Q_uv_dp_edge, Q_vv_dp_edge, &
+                                uvb_face_ave, H_bcl_edge, ope2_face_ave
+ 
+   implicit none
+ 
+   !global arrays
+   real, intent(inout) :: rhs(3,npoin)
+   real,intent(in):: q_send(nvarb+1,ngl,nboun)
+   real,intent(in):: q_recv(nvarb+1,ngl,nboun)
+   integer, intent(in) :: nvarb
+ 
+   !local variables
+ 
+   real :: wq, a_constant, iflux
+   integer ::iface,k
+   integer :: iel, ier, ilocl, ilocr
+   integer :: ifaceb, nq_i, nq_j, plane_ij
+ 
+   integer :: isub, ic, jj, im, imm, inbh, ib, kk, ivar
+   integer :: ftype, pface, subface, imulti
+   integer :: multirate
+   real, dimension(nvarb,nq) :: qbl, qbr
+   real, dimension(nq) :: pbl, pbr, H_face_temp
+   real :: nxl, nyl, nxr, nyr
+   real :: pbpert_edge, c_minus, c_plus
+   real :: c_pb_L, c_pb_R, c_pbub_LR, c_pbub_L, c_pbub_R
+   real, dimension(nq) :: one_plus_eta_edge, c_pb_LR
+   real, dimension(nq) :: flux_edge_x, flux_edge_y
+   integer :: el, iquad, n, I, il, jl, kl
+   real, dimension(nq) :: quu, quv, qvu, qvv, ul, ur, vl, vr
+   real :: pU_L, pU_R, lamb, dispu, dispv, flux_x, flux_y, flux
+   real :: hi, H_kx, H_ky
+ 
+   !Constants
+ 
+   jj=1
+   kk=1
+   imm = 0
+
+   ! do ifaceb =1,nboun
+   !    iface = face_send(ifaceb)  ! Get Local Face
+ 
+   do inbh = 1, num_nbh
+      do ib=1,num_send_recv(inbh)
+         iface = nbh_send_recv(jj)
+         imulti = nbh_send_recv_multi(jj)
+ 
+         ftype = face_type(iface)
+ 
+         do im = 1,imulti
+
+            el=face(7,iface)
+ 
[…]
