### 2025-11-29 67b5ce4
Update on the btp communication routines
