### 2025-11-30 b47f35b
Update on btp communications
