### 2025-11-30 4281496
Add elements face check in the communication routines
