+++ src/send_receive_bound.F90
@@ -36,12 +36,12 @@ subroutine unpack_data_dg_general_quad(q_send,q_recv,send_data,recv_data,nvarb)
         do ib = 1,num_send_recv(inbh)
             iface = nbh_send_recv(jj)
             imulti = nbh_send_recv_multi(jj)
-            !ftype = face_type(iface)
+            ftype = face_type(iface)
 
-            ! if (ftype == 2 .and. imulti>0) then
-                ilocl = face(5,iface)
+            if (ftype == 2 .and. imulti>0) then
+                ! ilocl = face(5,iface)
 
-                call mod_grid_get_face_nq(ilocl, nq_i, nq_j, plane_ij)
+                ! call mod_grid_get_face_nq(ilocl, nq_i, nq_j, plane_ij)
 
                 do inode = 1,nq
                     do ivar = 1,nvarb
@@ -51,7 +51,7 @@ subroutine unpack_data_dg_general_quad(q_send,q_recv,send_data,recv_data,nvarb)
                     end do
                 end do
                 kk=kk+1
-            ! end if
+            end if
             jj = jj + 1
         end do
     end do
@@ -87,10 +87,9 @@ subroutine unpack_data_dg_general_df(q_send,q_recv,send_data,recv_data,nvarb)
         do ib = 1,num_send_recv(inbh)
             iface = nbh_send_recv(jj)
             imulti = nbh_send_recv_multi(jj)
-            ftype = face_type(iface)
 
-            if (ftype == 2 .and. imulti>0) then
-                ilocl = face(5,iface)
+            if (face_type(iface) == 2 .and. imulti>0) then
+                ! ilocl = face(5,iface)
 
                 do inode = 1,ngl
                     do ivar = 1,nvarb+1
@@ -136,12 +135,11 @@ subroutine unpack_data_dg_general_quad_layer(q_send,q_recv,send_data,recv_data,n
         do ib = 1,num_send_recv(inbh)
             iface = nbh_send_recv(jj)
             imulti = nbh_send_recv_multi(jj)
-            !ftype = face_type(iface)
 
-            !if (ftype == 2 .and. imulti>0) then
-                ilocl = face(5,iface)
+            if (face_type(iface) == 2 .and. imulti>0) then
+                ! ilocl = face(5,iface)
 
-                call mod_grid_get_face_nq(ilocl, nq_i, nq_j, plane_ij)
+                ! call mod_grid_get_face_nq(ilocl, nq_i, nq_j, plane_ij)
 
                 do ll = 1,nlayers
 
@@ -154,7 +152,7 @@ subroutine unpack_data_dg_general_quad_layer(q_send,q_recv,send_data,recv_data,n
                     end do
                 end do
                 kk=kk+1
-            !end if
+            end if
             jj = jj + 1
         
         end do
@@ -190,12 +188,11 @@ subroutine unpack_data_dg_general_quad_1v(q_send,q_recv,send_data,recv_data)
         do ib = 1,num_send_recv(inbh)
             iface = nbh_send_recv(jj)
             imulti = nbh_send_recv_multi(jj)
-            !ftype = face_type(iface)
 
-            !if (ftype == 2 .and. imulti > 0) then
-                ilocl = face(5,iface)
+            if (face_type(iface) == 2 .and. imulti > 0) then
+                ! ilocl = face(5,iface)
 
-                call mod_grid_get_face_nq(ilocl, nq_i, nq_j, plane_ij)
+                ! call mod_grid_get_face_nq(ilocl, nq_i, nq_j, plane_ij)
 
                 do inode = 1,nq
                     ii = ii + 1
@@ -203,7 +200,7 @@ subroutine unpack_data_dg_general_quad_1v(q_send,q_recv,send_data,recv_data)
[…]
