### 2025-12-01 84217bc
Update on mod_lapacian_quad
