+++ src/mod_laplacian_quad.F90
@@ -29,7 +29,7 @@ module mod_laplacian_quad
 
     ! Compute the Laplacian barotropic horizontal viscosity term using the LDG method and
     ! using nodal points
-    subroutine btp_create_laplacian(rhs_lap,qb_df)
+    subroutine btp_create_laplacian_v0(rhs_lap,qb_df)
 
         ! Barotropic horizontal viscosity 
 
@@ -113,6 +113,101 @@ module mod_laplacian_quad
 
         ! Compute interface integral 
         call create_rhs_laplacian_flux(rhs_lap,graduv_face)
+
+        !RHS of viscosity 
+        rhs_lap(1,:) = visc_mlswe*massinv(:)*rhs_lap(1,:)
+        rhs_lap(2,:) = visc_mlswe*massinv(:)*rhs_lap(2,:)
+
+    end subroutine btp_create_laplacian_v0
+
+    subroutine btp_create_laplacian(rhs_lap,qb_df)
+
+        ! Barotropic horizontal viscosity 
+
+        use mod_variables, only: pbprime_visc, btp_dpp_graduv, graduvb_ave, graduvb_face_ave, &
+                                    btp_graduv_dpp_face
+
+        real, intent (out) :: rhs_lap(2,npoin)
+        real, dimension(4,npoin), intent(in) :: qb_df
+
+        real, dimension(2,npoin) :: Uk
+        real :: graduv_face(4,2,ngl,nface)
+        integer :: iface, il, jl, kl, ir, jr, kr, iel, ier, iquad, Iq, c_jump,k, ivar
+        real, dimension(4, npoin) :: graduv
+        real :: un, nx, ny, ul, ur, vl, vr, dpp
+
+        Uk(1,:) = qb_df(3,:)/qb_df(1,:)
+        Uk(2,:) = qb_df(4,:)/qb_df(1,:)
+
+        ! Compute the auxilary LDG variable graduv
+        call compute_gradient_uv(graduv, Uk)
+
+        graduvb_ave = graduvb_ave + graduv
+
+        ! Get the auxilary LDG variable graduv at the elements faces
+        do iface=1,nface
+
+            !Store Left and Right Side Variables
+            iel=face(7,iface)
+            ier=face(8,iface)
+    
+            !----------------------------Left Element
+            do iquad = 1,ngl
+
+                il=imapl(1,iquad,1,iface)
+                jl=imapl(2,iquad,1,iface)
+                kl=imapl(3,iquad,1,iface)
+                Iq = intma(il,jl,kl,iel)
+
+                graduv_face(:,1,iquad,iface) = graduv(:,Iq) 
+
+                if (ier > 0 ) then
+
+                    ir=imapr(1,iquad,1,iface)
+                    jr=imapr(2,iquad,1,iface)
+                    kr=imapr(3,iquad,1,iface)
+                    Iq=intma(ir,jr,kr,ier)
+
+                    graduv_face(:,2,iquad,iface) = graduv(:,Iq)
+
+                else
+                    graduv_face(:,2,iquad,iface) = graduv_face(:,1,iquad,iface)
+
+                    if(ier == -4) then 
+
+                        nx = normal_vector(1,iquad,1,iface)
+                        ny = normal_vector(2,iquad,1,iface)
+
+                        un = graduv_face(1,1,iquad,iface)*nx + graduv_face(2,1,iquad,iface)*ny
+                        graduv_face(1,2,iquad,iface) = graduv_face(1,1,iquad,iface) - 2.0*un*nx
+                        graduv_face(2,2,iquad,iface) = graduv_face(2,1,iquad,iface) - 2.0*un*ny
+
+                        un = graduv_face(3,1,iquad,iface)*nx + graduv_face(4,1,iquad,iface)*ny
[…]
