+++ src/create_rhs_dynamics_flux.F90
@@ -388,6 +388,123 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
  
  end subroutine create_nbhs_face_df_v1
 
+ subroutine create_nbhs_face_df_lap(rhs,q_send,q_recv,nvarb,multirate)
+
+   use mod_basis, only: ngl, FACE_CHILDREN,nq, psi
+ 
+   use mod_face, only: normal_vector, jac_face, imapl, imapr, face_send
+ 
+   use mod_grid, only: nelem, npoin, intma, face, nboun, face_type,nface
+ 
+   use mod_metrics, only: jac, massinv
+ 
+   use mod_parallel, only: nbh_send_recv, nbh_send_recv_multi, nbh_send_recv_half, &
+                           num_nbh, num_send_recv
+
+   use mod_variables, only: btp_graduv_dpp_face, graduvb_face_ave
+ 
+   implicit none
+ 
+   !global arrays
+   real, intent(inout) :: rhs(2,npoin)
+   real,intent(in):: q_send(nvarb+1,ngl,nboun)
+   real,intent(in):: q_recv(nvarb+1,ngl,nboun)
+   integer, intent(in) :: nvarb
+ 
+   !local variables
+ 
+   real :: wq, a_constant, iflux
+   integer ::iface,k
+   integer :: iel, ier, ilocl, ilocr
+   integer :: ifaceb, nq_i, nq_j, plane_ij
+ 
+   integer :: isub, ic, jj, im, imm, inbh, ib, kk, ivar
+   integer :: ftype, pface, subface, imulti
+   integer :: multirate
+   real :: nxl, nyl, nxr, nyr
+   integer :: el, iquad, n, I, il, jl, kl, er, itype, ip
+
+   real, dimension(2) :: qu_mean, qv_mean
+   real, dimension(4,2) :: flux_uv_visc_face
+   real, dimension(2) :: qul,qur
+   real, dimension(2) :: qvl,qvr
+   real :: flux_qu, flux_qv, hi, mul, mur,c_jump, alpha, beta
+
+   jj=1
+   kk=1
+   imm = 0
+
+   beta = 0.5
+   alpha = 1.0 - beta
+
+   do inbh = 1, num_nbh
+      do ib=1,num_send_recv(inbh)
+         iface = nbh_send_recv(jj)
+         imulti = nbh_send_recv_multi(jj)
+
+         el=face(7,iface)
+         er=face(8,iface)
+         
+         do iquad = 1,ngl
+
+            do ivar = 1,4
+               flux_uv_visc_face(ivar,1) = btp_graduv_dpp_face(5,1,iquad,iface)* &
+                                          q_send(ivar,iquad,kk) + &
+                                          btp_graduv_dpp_face(ivar,1,iquad,iface)
+                                       
+               flux_uv_visc_face(ivar,2) = btp_graduv_dpp_face(5,2,iquad,iface)* &
+                                          q_recv(ivar,iquad,kk) + &
+                                          btp_graduv_dpp_face(ivar,2,iquad,iface)
+
+               graduvb_face_ave(ivar,1,iquad,iface) = graduvb_face_ave(ivar,1,iquad,iface) + q_send(ivar,iquad,kk)
+               graduvb_face_ave(ivar,2,iquad,iface) = graduvb_face_ave(ivar,2,iquad,iface) + q_recv(ivar,iquad,kk)
+
+            end do
+
+            nxl = normal_vector(1,iquad,1,iface)
+            nyl = normal_vector(2,iquad,1,iface)
+
+            qul(:) = flux_uv_visc_face(1:2,1)
[…]
