### 2025-12-03 0b16fc9
Clean up the communications routines
