### 2025-12-03 f651421
Clean up the communications routines
