### 2025-12-05 fb42eba
Barotropic routines Crefactoring
