+++ src/mod_rhs_btp.F90
@@ -38,7 +38,7 @@ contains
 
         call create_rhs_btp_volume_qdf(rhs, qb_df, qprime_df)
 
-        call create_btp_fluxes_qdf(rhs,qb_df)
+        call create_btp_fluxes_qdf(rhs,qb_df, qprime_df)
 
         call btp_create_postcommunicator(rhs,4)
 
@@ -74,12 +74,12 @@ contains
         real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
         real, dimension(3,npoin), intent(out) :: rhs
 
-        real :: sc_x, sc_y, Hq, qu, quv, qv
+        real :: sc_x, sc_y, Hq, qu, quv, qvu, qv, H_b
         real :: wq, hi, dhdx, dhdy, coef_fric, tb_u, tb_v, ope, &
-                dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd
-        real :: pp, up, vp, pbq
-
-        integer :: I, Iq, ip
+                dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd, pbq
+        real, dimension(nlayers) :: pp, up, vp
+        real, dimension(nlayers+1) :: pprime
+        integer :: I, Iq, ip, k
 
         !speed1 = cd_mlswe/gravity
 
@@ -94,6 +94,7 @@ contains
 
             dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0
             pp = 0.0; up = 0.0; vp = 0.0; pbq = 0.0
+            pprime(1) = 0.0 ; H_b = 0.0
 
             do ip = 1,npts
                 I = indexq(ip,Iq)
@@ -103,24 +104,37 @@ contains
                 dpp = dpp + hi*qb_df(2,I)
                 udp = udp + hi*qb_df(3,I)
                 vdp = vdp + hi*qb_df(4,I)
-                pp = pp + hi*qprime_df(1,I,nlayers)
-                up = up + hi*qprime_df(2,I,nlayers)
-                vp = vp + hi*qprime_df(3,I,nlayers)
+                ! pp(:) = pp(:) + hi*qprime_df(1,I,:)
+                ! up(:) = up(:) + hi*qprime_df(2,I,:)
+                ! vp(:) = vp(:) + hi*qprime_df(3,I,:)
                 pbq = pbq + hi*pbprime_df(I)
             end do
+
+            do k = 1, nlayers
+                do ip = 1,npts
+                    I = indexq(ip,Iq)
+                    hi = psih(ip,Iq)
+                    pp(k) = pp(k) + hi*qprime_df(1,I,k)
+                    up(k) = up(k) + hi*qprime_df(2,I,k)
+                    vp(k) = vp(k) + hi*qprime_df(3,I,k)
+                enddo
+
+                pprime(k+1) = pprime(k) + pp(k)
+                H_b = H_b + 0.5*alpha_mlswe(k)*(pprime(k+1)**2 - pprime(k)**2)
+            enddo 
 
             wq = wjac(Iq)
             ub = udp/dp; vb = vdp/dp
 
             if (botfr == 1) then
-                ubot = up + ub
-                vbot = vp + vb
-                spd = (cd_mlswe/gravity)*pp
+                ubot = up(nlayers) + ub
+                vbot = vp(nlayers) + vb
+                spd = (cd_mlswe/gravity)*pp(nlayers)
                 tb_u = spd*ubot
                 tb_v = spd*vbot
             elseif (botfr == 2) then
[…]
