### 2025-12-05 c0f0817
Update on btp communication routines
