+++ src/send_receive_bound.F90
@@ -66,6 +66,54 @@ subroutine unpack_data_dg_general_df(q_send,q_recv,send_data,recv_data,nvarb)
 
     use mod_parallel, only: num_nbh, num_send_recv, nbh_send_recv, nbh_send_recv_multi
     
+    use mod_ref, only: nbtp_var
+
+    implicit none
+
+    !Global Variables
+    real, dimension(nbtp_var,ngl,nboun), intent(out) :: q_send,q_recv
+    real, dimension(nbtp_var*ngl*nboun), intent(in)  :: send_data, recv_data
+    integer, intent(in) :: nvarb
+
+    !Local Variables
+    integer ii, jj, kk, i, inbh, ib, ifaces, inode, jnode, ivar, ilocl, ilocr
+    integer nq_i, nq_j, plane_ij, iface, imulti, ftype
+
+    ii = 0
+    jj = 1
+    kk = 1
+    
+    do inbh = 1,num_nbh
+        do ib = 1,num_send_recv(inbh)
+            iface = nbh_send_recv(jj)
+            imulti = nbh_send_recv_multi(jj)
+
+            if (face_type(iface) == 2 .and. imulti>0) then
+                ! ilocl = face(5,iface)
+
+                do inode = 1,ngl
+                    do ivar = 1,nbtp_var
+                        ii = ii + 1
+                        q_send(ivar,inode,kk) = send_data(ii)
+                        q_recv(ivar,inode,kk) = recv_data(ii)
+                    end do
+                end do
+                kk=kk+1
+            end if
+            jj = jj + 1
+        end do
+    end do
+
+end subroutine unpack_data_dg_general_df
+
+subroutine unpack_data_dg_general_lap(q_send,q_recv,send_data,recv_data,nvarb)
+
+    use mod_basis, only: ngl, FACE_CHILDREN
+
+    use mod_grid, only: nboun, face, mod_grid_get_face_nq, face_type
+
+    use mod_parallel, only: num_nbh, num_send_recv, nbh_send_recv, nbh_send_recv_multi
+    
     !use mod_ref, only: nfields => nmessage
 
     implicit none
@@ -104,7 +152,7 @@ subroutine unpack_data_dg_general_df(q_send,q_recv,send_data,recv_data,nvarb)
         end do
     end do
 
-end subroutine unpack_data_dg_general_df
+end subroutine unpack_data_dg_general_lap
 
 subroutine unpack_data_dg_general_quad_layer(q_send,q_recv,send_data,recv_data,nvarb,nlayers,nq)
 
@@ -331,7 +379,84 @@ subroutine pack_data_dg_df(q_send,q_face,nvarb)
   
 end subroutine pack_data_dg_df
 
-subroutine pack_data_dg_df_btp(q_send,q,nvarb)
+subroutine pack_data_dg_df_btp(q_send,q,qprime_df,nvarb)
+  
+    use mod_basis, only: ngl, FACE_CHILDREN
+
+    use mod_face, only: face_send, imapl, imapr, normal_vector, jac_face
+
+    use mod_grid, only: nelem, npoin, intma, face_type, nboun, face, &
+                        mod_grid_get_face_nq, nface
+
+    use mod_initial, only: nvar, pbprime_df
+
+    use mod_metrics, only: jac
+
+    use mod_parallel, only: num_nbh, num_send_recv, nbh_send_recv, nbh_send_recv_multi
+
+    use mod_input, only: nlayers
+
+    use mod_ref, only: nbtp_var
+
+    implicit none
+  
+    !Global Variables
+    real, intent(out) :: q_send(nbtp_var*ngl*nboun)
+    real, intent(in) :: q(nvarb,npoin)
+    real, intent(in) :: qprime_df(3,npoin,nlayers)
[…]

[…]
