### 2025-12-05 64a9556
Barotropic communication code refactoring
