### 2025-12-06 f280a78
Update on btp/bcl laplacian communication routines
