+++ src/mod_laplacian_quad.F90
@@ -174,24 +174,27 @@ module mod_laplacian_quad
     ! using nodal points
     subroutine bcl_create_laplacian(rhs_lap)
 
-        use mod_variables, only: dpprime_visc, dpp_graduv, graduvb_ave
+        use mod_variables, only: dpprime_visc, dpp_graduv, graduvb_ave, dpprime_visc
 
         real, intent (out) :: rhs_lap(2,npoin,nlayers)
 
         integer :: k
-        real :: rhs_temp(2,npoin)
 
-        rhs_lap = 0.0
+        ! do k = 1,nlayers
+
+        call bcl_lap_create_precommunicator(dpp_graduv,dpprime_visc)
+
+        call bcl_compute_laplacian(rhs_lap)
+        call bcl_create_rhs_laplacian_flux(rhs_lap)
+
+        call bcl_create_rhs_lap_postcommunicator_df(rhs_lap)
 
         do k = 1,nlayers
-
-            call bcl_compute_laplacian(rhs_temp,k)
-            call bcl_create_rhs_laplacian_flux(rhs_temp, k)
-
-            rhs_lap(1,:,k) = visc_mlswe*massinv(:)*rhs_temp(1,:)
-            rhs_lap(2,:,k) = visc_mlswe*massinv(:)*rhs_temp(2,:)
-
+            rhs_lap(1,:,k) = visc_mlswe*massinv(:)*rhs_lap(1,:,k)
+            rhs_lap(2,:,k) = visc_mlswe*massinv(:)*rhs_lap(2,:,k)
         end do
+
+        ! end do
 
     end subroutine bcl_create_laplacian
 
@@ -337,36 +340,38 @@ module mod_laplacian_quad
 
     end subroutine btp_compute_laplacian
 
-    subroutine bcl_compute_laplacian(lap_q,k)
+    subroutine bcl_compute_laplacian(lap_q)
 
         use mod_variables, only: dpprime_visc, graduvb_ave, dpp_graduv
+        use mod_input, only: nlayers
 
         implicit none
 
-        real, intent(out) :: lap_q(2,npoin)
-        integer, intent(in) :: k
+        real, intent(out) :: lap_q(2,npoin,nlayers)
 
-        integer :: Iq, I, ip
+        integer :: Iq, I, ip, k
         real :: wq, qq(4)
-        
+
         lap_q = 0.0
 
-        ! Compute the gradient of q
-        do Iq = 1,npoin
+        do k = 1,nlayers
+            ! Compute the gradient of q
+            do Iq = 1,npoin
 
-            wq = wjac_df(Iq)
+                wq = wjac_df(Iq)
 
-            qq(1) = dpprime_visc(Iq,k)*graduvb_ave(1,Iq) + dpp_graduv(1,Iq,k)
-            qq(2) = dpprime_visc(Iq,k)*graduvb_ave(2,Iq) + dpp_graduv(2,Iq,k)
-            qq(3) = dpprime_visc(Iq,k)*graduvb_ave(3,Iq) + dpp_graduv(3,Iq,k)
-            qq(4) = dpprime_visc(Iq,k)*graduvb_ave(4,Iq) + dpp_graduv(4,Iq,k)
+                qq(1) = dpprime_visc(Iq,k)*graduvb_ave(1,Iq) + dpp_graduv(1,Iq,k)
+                qq(2) = dpprime_visc(Iq,k)*graduvb_ave(2,Iq) + dpp_graduv(2,Iq,k)
+                qq(3) = dpprime_visc(Iq,k)*graduvb_ave(3,Iq) + dpp_graduv(3,Iq,k)
[…]
