+++ output/run_script_serial.sh
@@ -2,5 +2,5 @@
 
 cp ../bin/numo3d .
 
-# mpirun -np 2 ./numo3d
-./numo3d
+mpirun -np 8 ./numo3d
+# ./numo3d
