+++ output/numo3d.in
@@ -35,8 +35,8 @@ time_scale = 1, !1=seconds, 3600=hours, 86400=days
 ad_mlswe = 0.0,
 botfr = 0, ! bottom Friction flag: 0 = no friction, 1 = linear, 2 = quadratic
 cd_mlswe = 0.0, ! cd = 1.0e-7 for linear, cd = 1e-3 (default) for quadratic bottom friction
-method_visc = 0,  ! 0=no viscosity
-visc_mlswe = 0.0, ! horizontal viscosity
+method_visc = 2,  ! 0=no viscosity
+visc_mlswe = 0.5, ! horizontal viscosity
 dg_integ_exact = T,
 dump_data = T,    ! write output files
 beta = 0.0,  ! beta-plane Coriolis parameter (default = 0.0)

[…]
