### 2025-12-07 1f7899f
blc communication routines
