+++ src/mod_create_rhs_mlswe.F90
@@ -21,7 +21,7 @@ module mod_create_rhs_mlswe
 
     public :: layer_momentum_rhs, &
               interpolate_layer_from_quad_to_node, rhs_layer_shear_stress, layer_mass_rhs, &
-              consistency_mass_rhs
+              consistency_mass_rhs, bcl_rhs
               
     contains
 
@@ -76,6 +76,36 @@ module mod_create_rhs_mlswe
         end do
         
     end subroutine layer_mass_rhs
+
+    subroutine bcl_rhs(rhs, rhs_visc, qprime_df, q_df, qprime_df_face)
+
+        use mod_input, only: nlayers
+        use mod_metrics, only: massinv
+        use mod_grid, only: npoin, npoin_q, intma, intma_dg_quad, nface
+        use mod_basis, only: nq, ngl
+
+        implicit none
+        real, dimension(3,npoin,nlayers), intent(out) :: rhs
+        real, dimension(2,npoin,nlayers), intent(in) :: rhs_visc
+        real, dimension(3,npoin,nlayers), intent(in) :: q_df, qprime_df
+        real, dimension(3,2,ngl,nface,nlayers), intent(in) :: qprime_df_face
+
+        integer :: k,I
+
+        call bcl_create_precommunicator(qprime_df)
+
+        call create_rhs_dynamics_volume_bcl(rhs, qprime_df, q_df)
+        call Apply_bcl_fluxes(rhs, qprime_df)
+
+        call bcl_create_postcommunicator(rhs)
+
+        do k = 1, nlayers
+            rhs(1,:,k) = massinv(:)*rhs(1,:,k)
+            rhs(2,:,k) = massinv(:)*rhs(2,:,k) + rhs_visc(1,:,k)
+            rhs(3,:,k) = massinv(:)*rhs(3,:,k) + rhs_visc(2,:,k)
+        end do
+
+    end subroutine bcl_rhs
 
     subroutine consistency_mass_rhs(dp_advec, qprime, flux_deficit_mass_face)
 
@@ -456,6 +486,194 @@ module mod_create_rhs_mlswe
 
     end subroutine create_rhs_dynamics_volume_layers
 
+    subroutine create_rhs_dynamics_volume_bcl(rhs, qprime_df, q_df)
+
+        use mod_grid, only : npoin_q, npoin, nelem, intma_dg_quad, intma
+        use mod_basis, only: npts, dpsiqx
+        use mod_input, only: nlayers
+        use mod_constants, only: gravity
+        use mod_initial, only: psih, dpsidx,dpsidy, indexq, wjac, alpha_mlswe, &
+                                tau_wind, pbprime_df, zbot_df
+        use mod_variables, only: tau_bot_ave, ope_ave, uvb_ave, H_ave, &
+                                    Qu_ave, Qv_ave, Quv_ave, uvb_ave, &
+                                    ope2_ave_df, ope2_ave, sum_layer_mass_flux
+
+        implicit none
+
+        real, dimension(3,npoin,nlayers), intent(out) :: rhs
+        real, dimension(3,npoin,nlayers), intent(in) :: q_df, qprime_df
+
+        real :: wq, hi, dhdx, dhdy, bot_layer, tau_wind_u, tau_wind_v, temp1
+        real :: Hq, var_uu, var_uv, var_vu, var_vv, source_x, source_y, Pstress
+        integer :: k, I, Iq, ip
+        real, dimension(nlayers+1) :: pprime_temp
+        real :: temp_dp, temp_u, temp_v, Ptop_k, Pbot_k, tempbot, Pbstress
+        real :: weight, acceleration, pbq
+        real, dimension(3) :: qp, qb
[…]
