+++ src/send_receive_bound.F90
@@ -154,6 +154,56 @@ subroutine unpack_data_dg_general_lap(q_send,q_recv,send_data,recv_data,nvarb)
 
 end subroutine unpack_data_dg_general_lap
 
+subroutine unpack_data_dg_general_bcl(q_send,q_recv,send_data,recv_data)
+
+    use mod_basis, only: ngl, FACE_CHILDREN
+
+    use mod_grid, only: nboun, face, mod_grid_get_face_nq, face_type
+
+    use mod_parallel, only: num_nbh, num_send_recv, nbh_send_recv, nbh_send_recv_multi
+    
+    use mod_input, only: nlayers
+
+    implicit none
+
+    !Global Variables
+    real, dimension(3*nlayers,ngl,nboun), intent(out) :: q_send,q_recv
+    real, dimension(3*nlayers*ngl*nboun), intent(in)  :: send_data, recv_data
+
+    !Local Variables
+    integer ii, jj, kk, i, inbh, ib, ifaces, inode, jnode, ivar, ilocl, ilocr
+    integer nq_i, nq_j, plane_ij, iface, imulti, ftype, index, ll
+
+    ii = 0
+    jj = 1
+    kk = 1
+    
+    do inbh = 1,num_nbh
+        do ib = 1,num_send_recv(inbh)
+            iface = nbh_send_recv(jj)
+            imulti = nbh_send_recv_multi(jj)
+
+            if (face_type(iface) == 2 .and. imulti>0) then
+                ! ilocl = face(5,iface)
+
+                do ll = 1,nlayers
+                    index = (ll-1)*3
+                    do inode = 1,ngl
+                        do ivar = 1,3
+                            ii = ii + 1
+                            q_send(index+ivar,inode,kk) = send_data(ii)
+                            q_recv(index+ivar,inode,kk) = recv_data(ii)
+                        end do
+                    end do
+                end do
+                kk=kk+1
+            end if
+            jj = jj + 1
+        end do
+    end do
+
+end subroutine unpack_data_dg_general_bcl
+
 subroutine unpack_data_dg_general_lap_bcl(q_send,q_recv,send_data,recv_data,nlayers)
 
     use mod_basis, only: ngl, FACE_CHILDREN
@@ -173,7 +223,7 @@ subroutine unpack_data_dg_general_lap_bcl(q_send,q_recv,send_data,recv_data,nlay
 
     !Local Variables
     integer ii, jj, kk, i, inbh, ib, ifaces, inode, jnode, ivar, ilocl, ilocr
-    integer nq_i, nq_j, plane_ij, iface, imulti, ftype, ll
+    integer nq_i, nq_j, plane_ij, iface, imulti, ftype, ll, index
 
     ii = 0
     jj = 1
@@ -188,11 +238,12 @@ subroutine unpack_data_dg_general_lap_bcl(q_send,q_recv,send_data,recv_data,nlay
                 ! ilocl = face(5,iface)
 
                 do ll = 1,nlayers
+                    index = (ll-1)*5
                     do inode = 1,ngl
                         do ivar = 1,5
                             ii = ii + 1
-                            q_send(ivar,inode,kk) = send_data(ii)
-                            q_recv(ivar,inode,kk) = recv_data(ii)
+                            q_send(index+ivar,inode,kk) = send_data(ii)
+                            q_recv(index+ivar,inode,kk) = recv_data(ii)
                         end do
                     end do
                 end do
@@ -580,7 +631,7 @@ subroutine pack_data_dg_df_btp_lap(q_send,q,btp_dpp_graduv,pbprime_visc,nvarb)
   
[…]
