### 2025-12-07 0a593a0
Refactor the communication and computation in momentum_mass routine
