+++ src/create_rhs_dynamics_flux.F90
@@ -504,6 +504,9 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
          iface = nbh_send_recv(jj)
          imulti = nbh_send_recv_multi(jj)
 
+         el=face(7,iface)
+         er=face(8,iface)
+
          qbl(1,:) = ope_face_ave(1,:,iface)
          qbl(2,:) = uvb_face_ave(1,1,:,iface)
          qbl(3,:) = uvb_face_ave(2,1,:,iface)
@@ -514,239 +517,239 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
          ql = 0.0; qr = 0.0
          do iquad = 1,nq
 
-               nxl = normal_vector_q(1,iquad,1,iface)
-               nyl = normal_vector_q(2,iquad,1,iface)
+            nxl = normal_vector_q(1,iquad,1,iface)
+            nyl = normal_vector_q(2,iquad,1,iface)
 
-               do k = 1,nlayers
+            do k = 1,nlayers
 
-                  index = 3*(k-1)
+               index = 3*(k-1)
 
-                  do n = 1, ngl
-                     hi = psiq(n,iquad)
-                     do ivar = 1,3
-                        !Left Element
-                        ql(ivar,iquad,k) = ql(ivar,iquad,k) + hi*q_send(index+ivar,n,kk)
-                        !Right Element
-                        qr(ivar,iquad,k) = qr(ivar,iquad,k) + hi*q_recv(index+ivar,n,kk)
-                     end do
-                  enddo
-
-                  ! Left side of the edge
-                  dpl = qbl(1,iquad) * ql(1,iquad,k)
-                  dpr = qbr(1,iquad) * qr(1,iquad,k)
-                  ul = ql(2,iquad,k)+qbl(2,iquad)
-                  ur = qr(2,iquad,k)+qbr(2,iquad)
-                  vl = ql(3,iquad,k)+qbl(3,iquad)
-                  vr = qr(3,iquad,k)+qbr(3,iquad)
-
-                  uu = 0.5*(ul+ur)
-                  vv = 0.5*(vl+vr)
-                  udpl(iquad,k) = ul*dpl
-                  udpr(iquad,k) = ur*dpr
-                  vdpl(iquad,k) = vl*dpl
-                  vdpr(iquad,k) = vr*dpr
-
-                  nxl = normal_vector_q(1,iquad,1,iface)
-                  nyl = normal_vector_q(2,iquad,1,iface)
-
-                  if(uu*nxl > 0.0) then
-                     udp_flux(1,iquad,k) = uu * dpl
-                     udp_flux(2,iquad,k) = uu * (ul*dpl)
-                     vdp_flux(2,iquad,k) = uu * (vl*dpl)
-                  else
-                     udp_flux(1,iquad,k) = uu * dpr
-                     udp_flux(2,iquad,k) = uu * (ur*dpr)
-                     vdp_flux(2,iquad,k) = uu * (vr*dpr)
-                  endif
-                  if(vv*nyl > 0.0) then
-                     vdp_flux(1,iquad,k) = vv * dpl
-                     udp_flux(3,iquad,k) = vv * (ul*dpl)
-                     vdp_flux(3,iquad,k) = vv * (vl*dpl)
-                  else
-                     vdp_flux(1,iquad,k) = vv * dpr
-                     udp_flux(3,iquad,k) = vv * (ur*dpr)
-                     vdp_flux(3,iquad,k) = vv * (vr*dpr)
-                  endif
-
+               do n = 1, ngl
+                  hi = psiq(n,iquad)
+                  do ivar = 1,3
+                     !Left Element
[…]
