### 2025-12-07 5027897
Update on blc continuity communication routines
