+++ src/mod_create_rhs_mlswe.F90
@@ -65,11 +65,14 @@ module mod_create_rhs_mlswe
         
         integer :: k
 
+        call bcl_create_precommunicator(qprime_df)
         ! Compute the mass advection term for the degree of freedom for dp in each layer
         call create_layers_volume_mass(dp_advec, qprime_df)
 
         ! Compute the mass flux term 
-        call create_layer_mass_flux(dp_advec, qprime_df_face)
+        call create_layer_mass_flux(dp_advec, qprime_df)
+
+        call bcl_create_postcommunicator_continuity(dp_advec)
 
         do k = 1, nlayers
             dp_advec(:,k) = massinv(:)*dp_advec(:,k)
@@ -1051,6 +1054,7 @@ module mod_create_rhs_mlswe
         use mod_face, only: imapl, imapr, normal_vector_q, jac_faceq
         use mod_variables, only: ope_face_ave, H_face_ave, one_plus_eta_edge_2_ave, &
                                 uvb_face_ave, Quv_face_ave, Qu_face_ave, Qv_face_ave, ope2_face_ave
+        use mod_variables, only: sum_layer_mass_flux_face
 
         implicit none
 
@@ -1136,7 +1140,7 @@ module mod_create_rhs_mlswe
                             qr(3,iquad,k) = ql(3,iquad,k) - 2.0*un*nyl
                         elseif (er == -2) then
                             qr(2,iquad,k) = -ql(2,iquad,k)
-                            qr(3,iquad,k) = ql(3,iquad,k)
+                            qr(3,iquad,k) = -ql(3,iquad,k)
                         endif
                     end if
 
@@ -1154,9 +1158,6 @@ module mod_create_rhs_mlswe
                     udpr(iquad,k) = ur*dpr
                     vdpl(iquad,k) = vl*dpl
                     vdpr(iquad,k) = vr*dpr
-
-                    nxl = normal_vector_q(1,iquad,1,iface)
-                    nyl = normal_vector_q(2,iquad,1,iface)
 
                     if(uu*nxl > 0.0) then
                         udp_flux(1,iquad,k) = uu * dpl
@@ -1179,13 +1180,15 @@ module mod_create_rhs_mlswe
 
                 enddo
 
+                sum_layer_mass_flux_face(1,iquad,iface) = sum_layer_mass_flux_face(1,iquad,iface) &
+                                                        + udp_flux(1,iquad,iface)
+                sum_layer_mass_flux_face(2,iquad,iface) = sum_layer_mass_flux_face(2,iquad,iface) &
+                                                        + vdp_flux(1,iquad,iface)
+
                 uu_dp_flux_deficit = Qu_face_ave(1,iquad,iface) - sum(udp_flux(2,iquad,:))
                 uv_dp_flux_deficit = Qu_face_ave(2,iquad,iface) - sum(udp_flux(3,iquad,:))
                 vu_dp_flux_deficit = Qv_face_ave(1,iquad,iface) - sum(vdp_flux(2,iquad,:))
                 vv_dp_flux_deficit = Qv_face_ave(2,iquad,iface) - sum(vdp_flux(3,iquad,:))
-
-                nxl = normal_vector_q(1,iquad,1,iface)
-                nyl = normal_vector_q(2,iquad,1,iface)
 
                 ! Adjust the fluxes for the u-momentum equation
                 one_over_sum_l = 1.0 / sum(abs(udpl(iquad,:))+eps1)
@@ -1541,10 +1544,10 @@ module mod_create_rhs_mlswe
 
     end subroutine create_consistency_volume_mass
 
-    subroutine create_layer_mass_flux(dp_advec, qprime_df_face)
[…]
