+++ src/create_rhs_dynamics_flux.F90
@@ -458,6 +458,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
    use mod_initial, only : alpha_mlswe, zbot_face
    use mod_variables, only: ope_face_ave, H_face_ave, one_plus_eta_edge_2_ave, &
                                 uvb_face_ave, Quv_face_ave, Qu_face_ave, Qv_face_ave, ope2_face_ave
+   use mod_variables, only: sum_layer_mass_flux_face
  
    implicit none
  
@@ -549,9 +550,6 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                vdpl(iquad,k) = vl*dpl
                vdpr(iquad,k) = vr*dpr
 
-               nxl = normal_vector_q(1,iquad,1,iface)
-               nyl = normal_vector_q(2,iquad,1,iface)
-
                if(uu*nxl > 0.0) then
                   udp_flux(1,iquad,k) = uu * dpl
                   udp_flux(2,iquad,k) = uu * (ul*dpl)
@@ -571,15 +569,17 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                   vdp_flux(3,iquad,k) = vv * (vr*dpr)
                endif
 
-            enddo
+            enddo !k
+
+            sum_layer_mass_flux_face(1,iquad,iface) = sum_layer_mass_flux_face(1,iquad,iface)  + &
+                                                      sum(udp_flux(1,iquad,:))
+            sum_layer_mass_flux_face(2,iquad,iface) = sum_layer_mass_flux_face(2,iquad,iface)  + &
+                                                      sum(vdp_flux(1,iquad,:))
 
             uu_dp_flux_deficit = Qu_face_ave(1,iquad,iface) - sum(udp_flux(2,iquad,:))
             uv_dp_flux_deficit = Qu_face_ave(2,iquad,iface) - sum(udp_flux(3,iquad,:))
             vu_dp_flux_deficit = Qv_face_ave(1,iquad,iface) - sum(vdp_flux(2,iquad,:))
             vv_dp_flux_deficit = Qv_face_ave(2,iquad,iface) - sum(vdp_flux(3,iquad,:))
-
-            nxl = normal_vector_q(1,iquad,1,iface)
-            nyl = normal_vector_q(2,iquad,1,iface)
 
             ! Adjust the fluxes for the u-momentum equation
             one_over_sum_l = 1.0 / sum(abs(udpl(iquad,:))+eps1)
@@ -782,7 +782,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                   rhs(2,I,k) = rhs(2,I,k) - wq*hi*(hlx_k + flux_x)
                   rhs(3,I,k) = rhs(3,I,k) - wq*hi*(hly_k + flux_y)
                end do
-               end do
+            end do
          end do
 
          kk=kk+1
@@ -792,6 +792,136 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
    end do !iface
  
  end subroutine create_nbhs_face_bcl
+
+  subroutine create_nbhs_face_bcl_continuity(rhs,q_send,q_recv,multirate)
+
+   use mod_basis, only: ngl, FACE_CHILDREN,nq, psi, psiq
+   use mod_face, only: normal_vector_q, jac_faceq, imapl, imapr, face_send
+   use mod_grid, only: nelem, npoin, intma, face, nboun, face_type,nface
+   use mod_parallel, only: nbh_send_recv, nbh_send_recv_multi, nbh_send_recv_half, &
[…]
