+++ Makefile
@@ -2,7 +2,7 @@ RM = /bin/rm -rf
 MV = /bin/mv -f
 CP = /bin/cp -f
 
-export PLATFORMS:= hnumo hnumo-debug hnumo-brew hnumo-prof hnumo-CI hnumo-debian hnumo-debian-debug
+export PLATFORMS:= hnumo hnumo-debug hnumo-brew hnumo-brew-dbg hnumo-prof hnumo-CI hnumo-debian hnumo-debian-debug
 
 export NUMO_DIR    := $(CURDIR)
 export DEPEND_FILE := $(NUMO_DIR)/depend.mk
