### 2025-12-07 91ed639
Update on bcl communication routines
