+++ src/create_rhs_dynamics_flux.F90
@@ -685,12 +685,12 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                   dz_intersect = z_intersect_top - z_intersect_bot
 
                   if (dz_intersect > 0.0) then
-                        p_intersect_bot = p_edge_minus(ktemp+1) &
-                              - g_over_alpha(ktemp)*(z_intersect_bot - z_edge_minus(ktemp+1))
-                        p_intersect_top = p_edge_minus(ktemp+1) &
-                              - g_over_alpha(ktemp)*(z_intersect_top - z_edge_minus(ktemp+1))
-                        H_r_minus = H_r_minus + &
-                              0.5*alpha_mlswe(ktemp)*(p_intersect_bot**2 - p_intersect_top**2)
+                     p_intersect_bot = p_edge_minus(ktemp+1) &
+                           - g_over_alpha(ktemp)*(z_intersect_bot - z_edge_minus(ktemp+1))
+                     p_intersect_top = p_edge_minus(ktemp+1) &
+                           - g_over_alpha(ktemp)*(z_intersect_top - z_edge_minus(ktemp+1))
+                     H_r_minus = H_r_minus + &
+                           0.5*alpha_mlswe(ktemp)*(p_intersect_bot**2 - p_intersect_top**2)
 
                   end if
                end do
@@ -707,12 +707,12 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                   dz_intersect = z_intersect_top - z_intersect_bot
 
                   if (dz_intersect > 0.0) then
-                        p_intersect_bot = p_edge_plus(ktemp+1) &
-                           - g_over_alpha(ktemp)*(z_intersect_bot - z_edge_plus(ktemp+1))
-                        p_intersect_top = p_edge_plus(ktemp+1) &
-                           - g_over_alpha(ktemp)*(z_intersect_top - z_edge_plus(ktemp+1))
-                        H_r_plus = H_r_plus &
-                           + 0.5*alpha_mlswe(ktemp)*(p_intersect_bot**2 - p_intersect_top**2)
+                     p_intersect_bot = p_edge_plus(ktemp+1) &
+                        - g_over_alpha(ktemp)*(z_intersect_bot - z_edge_plus(ktemp+1))
+                     p_intersect_top = p_edge_plus(ktemp+1) &
+                        - g_over_alpha(ktemp)*(z_intersect_top - z_edge_plus(ktemp+1))
+                     H_r_plus = H_r_plus &
+                        + 0.5*alpha_mlswe(ktemp)*(p_intersect_bot**2 - p_intersect_top**2)
                   end if
                end do
                H_face(2,iquad,k) = 0.5*(H_r_plus + H_r_minus) ! H_r for the right side
@@ -922,6 +922,342 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
    end do !iface
  
  end subroutine create_nbhs_face_bcl_continuity
+
+ subroutine create_nbhs_face_bcl_momentum(rhs,q_send,q_recv,multirate)
+
+   use mod_basis, only: ngl, FACE_CHILDREN,nq, psi, psiq
+   use mod_face, only: normal_vector_q, jac_faceq, imapl, imapr, face_send
+   use mod_grid, only: nelem, npoin, intma, face, nboun, face_type,nface
+   use mod_metrics, only: jac, massinv
[…]
