+++ src/mod_create_rhs_mlswe.F90
@@ -40,8 +40,13 @@ module mod_create_rhs_mlswe
 
         integer :: k,I
 
+        call bcl_create_precommunicator(qprime_df)
+
         call create_rhs_dynamics_volume_layers(rhs_mom, qprime_df, q_df)
-        call Apply_layers_fluxes(rhs_mom, qprime_df_face)
+
+        call Apply_layers_fluxes(rhs_mom, qprime_df)
+
+        call bcl_create_postcommunicator_momentum(rhs_mom)
 
         do k = 1, nlayers
             rhs_mom(1,:,k) = massinv(:)*rhs_mom(1,:,k) + rhs_visc(1,:,k)
@@ -677,14 +682,14 @@ module mod_create_rhs_mlswe
 
     end subroutine create_rhs_dynamics_volume_bcl
 
-    subroutine Apply_layers_fluxes(rhs_mom, qprime_df_face)
+    subroutine Apply_layers_fluxes(rhs_mom, qprime_df)
 
         ! This routine computes the layer momentum advection flux terms using upwind flux
         ! This routine computes the layer momentum pressure terms
 
         use mod_constants, only : gravity
         use mod_initial, only : alpha_mlswe, zbot_face
-        use mod_grid, only : nface, npoin, npoin_q, face, intma
+        use mod_grid, only : nface, npoin, npoin_q, face, intma, face_type
         use mod_basis, only : nq, psiq, ngl
         use mod_input, only : nlayers
         use mod_face, only: imapl, imapr, normal_vector_q, jac_faceq
@@ -694,7 +699,7 @@ module mod_create_rhs_mlswe
         implicit none
 
         real, dimension(2, npoin, nlayers), intent(inout) :: rhs_mom
-        real, dimension(3,2,ngl,nface,nlayers), intent(in) :: qprime_df_face
+        real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
 
         real, dimension(nlayers) :: alpha_over_g, g_over_alpha
         real, dimension(2,nlayers+1) :: p_face, z_face
@@ -712,7 +717,7 @@ module mod_create_rhs_mlswe
         real :: vu_dp_flux_deficit, vv_dp_flux_deficit
         real, parameter :: eps1 = 1.0e-20 !  Parameter used to prevent division by zero.
         integer :: il, jl, ir, jr, kl, kr, jquad, n, m
-        real :: wq, hi, hlx_k, hly_k, hrx_k, hry_k, flux_x, flux_y, hx_k, hy_k
+        real :: wq, hi, hlx_k, hly_k, hrx_k, hry_k, flux_x, flux_y, hx_k, hy_k, un
         real, dimension(2,nq,nlayers) :: H_face, udp_flux, vdp_flux
 
         do k=1,nlayers
@@ -722,6 +727,8 @@ module mod_create_rhs_mlswe
 
         ! Compute H_r at the element face
         do iface = 1, nface
+
+            if (face_type(iface) == 2) cycle
 
             ! Store Left Side Variables
             el = face(7,iface)
@@ -743,10 +750,39 @@ module mod_create_rhs_mlswe
                 do k = 1,nlayers
 
                     do n = 1, ngl
+
+                        il = imapl(1,n,1,iface)
+                        jl = imapl(2,n,1,iface)
+                        kl = imapl(3,n,1,iface)
+                        I = intma(il,jl,kl,el)
                         hi = psiq(n,iquad)
-                        ql(:,iquad,k) = ql(:,iquad,k) + hi*qprime_df_face(:,1,n,iface,k)
-                        qr(:,iquad,k) = qr(:,iquad,k) + hi*qprime_df_face(:,2,n,iface,k)
+
[…]

[…]
