### 2025-12-07 a054a10
Rewrite the consistency computation/communication routines
