### 2025-12-07 a216158
Rewrite the btp/bcl communication routines
