+++ src/mod_splitting.F90
@@ -22,7 +22,7 @@ module mod_splitting
     
     contains
 
-    subroutine thickness(qprime_df, q_df, qb_df, qprime_df_face)
+    subroutine thickness(qprime_df, q_df, qb_df)
 
         ! ========================================================================================
         ! This subroutine is used to predict or correct the layer thickness for the splitting 
@@ -44,7 +44,6 @@ module mod_splitting
     
         ! Input variables
         real, dimension(3,npoin,nlayers), intent(inout) :: qprime_df
-        real, dimension(3,2,ngl,nface,nlayers), intent(inout) :: qprime_df_face
         real, dimension(3,npoin,nlayers), intent(inout) :: q_df
         real, intent(in)    :: qb_df(4,npoin)
     
@@ -57,7 +56,7 @@ module mod_splitting
         ! Compute the mass advection term and RHS for mass equation, 
         ! return the result in array  dp_advec. 
 
-        call layer_mass_rhs(dp_advec, qprime_df, qprime_df_face)
+        call layer_mass_rhs(dp_advec, qprime_df)
 
         ! Compute the tentative values of the predicted or corrected 
         ! degrees of freedom for  dp (q_df(1,:,:)).  These would be the values at
@@ -85,13 +84,11 @@ module mod_splitting
         do k = 1,nlayers
             qprime_df(1,:,k) = q_df(1,:,k) / one_plus_eta_temp(:)
         end do
-
-        call extract_dprime_df_face(qprime_df_face(1,:,:,:,:), qprime_df(1,:,:))
         
     end subroutine thickness
 
 
-    subroutine momentum(q_df,qprime_df,qb_df,qprime_df_face)
+    subroutine momentum(q_df,qprime_df,qb_df)
 
         ! ========================================================================================
         ! This subroutine is used to correct the layer momentum for the splitting system 
@@ -111,7 +108,6 @@ module mod_splitting
 
         ! Input variables
         real, dimension(3,npoin,nlayers), intent(inout) :: q_df
-        real, dimension(3,2,ngl,nface,nlayers), intent(in) :: qprime_df_face
         real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
         real, dimension(4,npoin), intent(in) :: qb_df
 
@@ -128,7 +124,7 @@ module mod_splitting
         q_df_temp = 0.0
 
         ! ==== layer momentum ====
-        call rhs_momentum(rhs_mom, qprime_df,q_df, qprime_df_face)
+        call rhs_momentum(rhs_mom, qprime_df,q_df)
 
         ! Compute the momentum equation variables for the next time step
         do k = 1,nlayers
@@ -184,7 +180,7 @@ module mod_splitting
 
     end subroutine momentum
 
-    subroutine momentum_mass(q_df,qprime_df_face,qprime_df,qb_df)
+    subroutine momentum_mass(q_df, qprime_df, qb_df)
 
         ! ========================================================================================
         ! This subroutine is used to predict the layer mass and momentum for the splitting system
@@ -196,7 +192,7 @@ module mod_splitting
         use mod_basis, only: nq, ngl
         use mod_input, only: nlayers, dt, ad_mlswe
         use mod_initial, only: fdt_bcl, fdt2_bcl, a_bcl, b_bcl, pbprime_df
-        use mod_create_rhs_mlswe, only: rhs_layer_shear_stress, layer_mass_rhs
+        use mod_create_rhs_mlswe, only: rhs_layer_shear_stress
         use mod_layer_terms, only: layer_mom_boundary_df, &
                                     velocity_df, extract_qprime_df_face,extract_velocity
[…]
