### 2025-12-13 1aef1e0
Working dry state limiter
