### 2026-01-04 c930cb0
Working on rk method for bcl solver
