+++ src/mod_variables.F90
@@ -26,11 +26,12 @@ module mod_variables
     real, dimension(:,:,:), allocatable :: Qu_face, Qv_face, one_plus_eta_face, flux_edge
     real, dimension(:,:), allocatable :: Q_uu_dp_edge, Q_uv_dp_edge, Q_vv_dp_edge, H_bcl_edge
     real, dimension(:), allocatable :: Quu, Qvv, Quv, H, one_plus_eta, one_plus_eta_df, ope2_ave_df
-    real, dimension(:), allocatable :: one_plus_eta_out, pbprime_visc, Qv_ave, Quv_ave, ope2_ave
-    real, dimension(:), allocatable :: Q_uu_dp, Q_uv_dp, Q_vv_dp, H_bcl, ope_ave, H_ave, Qu_ave 
+    real, dimension(:), allocatable :: one_plus_eta_out, pbprime_visc, ope2_ave
+    real, dimension(:), allocatable :: Q_uu_dp, Q_uv_dp, Q_vv_dp, H_bcl, ope_ave, H_ave 
     real, dimension(:,:), allocatable :: tau_bot, btp_mass_flux, H_face, one_plus_eta_edge_2
     real, dimension(:,:), allocatable :: btp_mass_flux_ave, uvb_ave, uvb_ave_df
     real, dimension(:,:), allocatable :: dpprime_visc, dpprime_visc_q
+    real, dimension(:,:), allocatable :: Qu_ave, Qv_ave
 
     real, dimension(:,:), allocatable :: one_plus_eta_edge_2_ave, H_face_ave, tau_wind_ave
     real, dimension(:,:), allocatable :: tau_bot_ave, one_plus_eta_edge
@@ -54,7 +55,7 @@ module mod_variables
         integer AllocateStatus
 
         if(allocated(H_bcl)) then 
-            deallocate(Q_uu_dp, Q_uv_dp, Q_vv_dp, H_bcl, ope_ave, H_ave, Qu_ave, Qv_ave, Quv_ave, &
+            deallocate(Q_uu_dp, Q_uv_dp, Q_vv_dp, H_bcl, ope_ave, H_ave, Qu_ave, Qv_ave, &
             ope2_ave, Quu, Qvv, Quv, H, one_plus_eta, one_plus_eta_df, ope2_ave_df, &
             one_plus_eta_out, Q_uu_dp_edge, Q_uv_dp_edge, Q_vv_dp_edge, H_bcl_edge, &
             Qu_face, Qv_face, one_plus_eta_face, flux_edge, &
@@ -70,8 +71,8 @@ module mod_variables
             one_plus_eta_face(2,nq,nface), flux_edge(2,nq,nface), one_plus_eta_df(npoin), &
             tau_bot(2,npoin_q), btp_mass_flux(2,npoin_q), H_face(nq,nface), &
             one_plus_eta_edge_2(nq,nface), Quu(npoin_q), Qvv(npoin_q), Quv(npoin_q), &
-            H(npoin_q), one_plus_eta(npoin_q), ope_ave(npoin_q), H_ave(npoin_q), Qu_ave(npoin_q), &
-            Qv_ave(npoin_q), Quv_ave(npoin_q), ope2_ave(npoin_q), btp_mass_flux_ave(2,npoin_q), &
+            H(npoin_q), one_plus_eta(npoin_q), ope_ave(npoin_q), H_ave(npoin_q), Qu_ave(2,npoin_q), &
+            Qv_ave(2,npoin_q), ope2_ave(npoin_q), btp_mass_flux_ave(2,npoin_q), &
             uvb_ave(2,npoin_q), ope2_ave_df(npoin), uvb_face_ave(2,2,nq,nface), &
[…]
