+++ output/run_script_serial.sh
@@ -2,5 +2,5 @@
 
 cp ../bin/numo3d .
 
-mpirun -np 4 ./numo3d
-# ./numo3d
+# mpirun -np 4 ./numo3d
+./numo3d
