### 2026-01-05 076fac9
Working on vanishing layers
