+++ src/mod_rhs_btp.F90
@@ -65,7 +65,7 @@ contains
                                 pbprime_df, coriolis_quad, alpha_mlswe
         use mod_variables, only: tau_bot_ave, H_ave, Qu_ave, Quv_ave, Qv_ave, ope_ave, &
                                 uvb_ave, btp_mass_flux_ave, ope2_ave
-        use mod_input, only: nlayers, cd_mlswe, botfr
+        use mod_input, only: nlayers, cd_mlswe, botfr, dry_cutoff
 
         implicit none
 
@@ -73,13 +73,12 @@ contains
         real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
         real, dimension(3,npoin), intent(out) :: rhs
 
-        real :: sc_x, sc_y, Hq, qu, quv, qvu, qv, H_b
+        real :: sc_x, sc_y, Hq, qu, quv, qvu, qv, H_b, pb_cutoff, pk, uk, vk
         real :: wq, hi, dhdx, dhdy, coef_fric, tb_u, tb_v, ope, &
-                dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd, pbq
+                dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd, pbq, H_bclq
         real, dimension(nlayers) :: pp, up, vp
         real, dimension(nlayers+1) :: pprime
         integer :: I, Iq, ip, k
-        real, parameter :: eps = 1.0e-6
 
         !speed1 = cd_mlswe/gravity
 
@@ -90,11 +89,13 @@ contains
         Pbstress = (gravity/alpha_mlswe(nlayers)) * 10.0 ! pressure corresponding to 10m depth
                                                          ! at which bottom stress is reduced to 0
 
+        pb_cutoff = (gravity/alpha_mlswe(nlayers))*dry_cutoff
+
         do Iq = 1, npoin_q
 
             dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0
             pp = 0.0; up = 0.0; vp = 0.0; pbq = 0.0
-            pprime(1) = 0.0 ; H_b = 0.0
+            pprime(:) = 0.0 ; H_b = 0.0
 
             do ip = 1,npts
                 I = indexq(ip,Iq)
@@ -107,10 +108,10 @@ contains
                 pbq = pbq + hi*pbprime_df(I)
             end do
 
-            if (dp <= (gravity/alpha_mlswe(nlayers))*eps) then
-                dp = (gravity/alpha_mlswe(nlayers))*eps
-                udp = 0.0 ; vdp = 0.0 ; dpp = 0.0
-            end if
+            ! if (dp <= pb_cutoff) then
+            !     dp = pb_cutoff
+            !     udp = 0.0 ; vdp = 0.0 ; dpp = 0.0
+            ! end if
 
             do k = 1, nlayers
                 do ip = 1,npts
@@ -121,14 +122,18 @@ contains
                     vp(k) = vp(k) + hi*qprime_df(3,I,k)
                 enddo
 
-                if (pp(k) <= (gravity/alpha_mlswe(k))*eps) then
-                    pp(k) = (gravity/alpha_mlswe(k))*eps
+                if (pp(k) <= (gravity/alpha_mlswe(k))*dry_cutoff) then
+                    pp(k) = (gravity/alpha_mlswe(k))*dry_cutoff
                     up(k) = 0.0 ; vp(k) = 0.0
                 end if
 
                 pprime(k+1) = pprime(k) + pp(k)
[…]
