+++ src/mod_input.F90
@@ -41,7 +41,7 @@ module mod_input
         ad_mlswe, cd_mlswe, &
         dp_tau_bot, dp_tau_wind, dt_btp, method_visc,visc_mlswe, max_shear_dz, &
         adjust_H_vertical_sum, botfr, dg_integ_exact, dump_data, lcheck_conserved, adjust_bcl_mom_flux, &
-        f0, beta
+        f0, beta, dry_cutoff
  
    public :: eqn_set, is_mlswe
  
@@ -127,6 +127,7 @@ module mod_input
    real(kind=r8) :: max_shear_dz = 0.0
    real(kind=r8) :: f0 = 0.0
    real(kind=r8) :: beta = 0.0
+   real(kind=r8) :: dry_cutoff = 1.0e-10
  
    !-----------------------------------------------------------------------
    ! Namelist Variables
@@ -346,7 +347,7 @@ module mod_input
          ad_mlswe, cd_mlswe, dp_tau_bot, dp_tau_wind, dt_btp,method_visc,&
          visc_mlswe, max_shear_dz, adjust_H_vertical_sum, botfr, &
          dg_integ_exact, dump_data, lcheck_conserved, adjust_bcl_mom_flux, &
-         f0, beta
+         f0, beta, dry_cutoff
  
      namelist /gridnl/ nelx, nely, nelz, nopx, nopy, nopz, xdims, ydims, ztop, zbottom, &
           nlayers, &
