+++ src/mod_create_rhs_mlswe.F90
@@ -320,7 +320,7 @@ module mod_create_rhs_mlswe
 
         use mod_grid, only : npoin_q, npoin, nelem, intma_dg_quad, intma
         use mod_basis, only: npts, dpsiqx
-        use mod_input, only: nlayers
+        use mod_input, only: nlayers, dry_cutoff
         use mod_constants, only: gravity
         use mod_initial, only: psih, dpsidx,dpsidy, indexq, wjac, alpha_mlswe, &
                                 tau_wind, pbprime_df, zbot_df
@@ -344,7 +344,7 @@ module mod_create_rhs_mlswe
         real :: p_tmp(nlayers+1), u, v, dp, weightq, one_over_sumuq, one_over_sumvq
         real :: uu_dp_deficitq, uv_dp_deficitq, vv_dp_deficitq, gradz(2,nlayers+1)
         real :: z_elv(npoin,nlayers+1)
-        real, parameter :: eps1 = 1.0e-20, eps = 1.0e-10 !  Parameter used to prevent division by zero.
+        real, parameter :: eps1 = 1.0e-20!  Parameter used to prevent division by zero.
 
         rhs_mom = 0.0
         bot_layer = 0.0
@@ -377,11 +377,11 @@ module mod_create_rhs_mlswe
                     temp_vv(k) = temp_vv(k) + hi*q_df(3,I,k)
                 enddo
 
-                if (qp(1) <= (gravity/alpha_mlswe(k))*eps) then
-                    qp(1) = (gravity/alpha_mlswe(k))*eps
-                    qp(2:3) = 0.0
-                    temp_uu(k) = 0.0 ; temp_vv(k) = 0.0
-                end if
+                ! if (qp(1) <= (gravity/alpha_mlswe(k))*dry_cutoff) then
+                !     qp(1) = (gravity/alpha_mlswe(k))*dry_cutoff
+                !     qp(2:3) = 0.0
+                !     temp_uu(k) = 0.0 ; temp_vv(k) = 0.0
+                ! end if
 
                 qb(1) = ope_ave(Iq)
                 qb(2) = uvb_ave(1,Iq)
@@ -389,10 +389,17 @@ module mod_create_rhs_mlswe
 
                 p_tmp(k+1) = p_tmp(k) + sqrt(ope2_ave(Iq)) * qp(1)
                 H_tmp(k) = 0.5*alpha_mlswe(k) * (p_tmp(k+1)**2 - p_tmp(k)**2)
+                if (abs(p_tmp(k+1) - p_tmp(k)) <= (gravity/alpha_mlswe(k))*dry_cutoff) H_tmp(k) = 0.0
 
                 dp = qp(1) * qb(1)
                 u = qp(2) + qb(2)
                 v = qp(3) + qb(3)
+
+                if (qp(1) <= (gravity/alpha_mlswe(k))*dry_cutoff) then
+                    dp = ((gravity/alpha_mlswe(k))*dry_cutoff)*qb(1)
+                    u = 0.0 ; v = 0.0
+                    temp_uu(k) = 0.0 ; temp_vv(k) = 0.0
+                end if
 
                 u_udp(k) = dp * u*u
                 v_vdp(k) = dp * v*v
@@ -479,6 +486,11 @@ module mod_create_rhs_mlswe
                 source_y = gravity*(tau_wind_v - tempbot*tau_bot_ave(2,Iq) + &
                             p_tmp(k) * gradz(2,k) - p_tmp(k+1) * gradz(2,k+1))
 
+                if (abs(p_tmp(k+1) - p_tmp(k)) <= (gravity/alpha_mlswe(k))*dry_cutoff) then
[…]

[…]
