### 2026-01-06 606f65a
Incorporate feedback from JOSS review
