+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'h-NUMO: A Hydrostatic Version of the Non-hydrostatic Model of the Ocean'
+title: 'h-NUMO: A high-performance discontinuous-Galerkin code for hydrostatic ocean modelling'
 tags:
   - isopycnal model
   - Ocean modeling
@@ -22,24 +22,25 @@ bibliography: paper.bib
 
 # Summary
 
-h-NUMO is a hydrostatic version of the Non-hydrostatic Model of the Ocean (NUMO). It is designed for large-scale ocean applications while maintaining computational efficiency. The model is implemented in the Galerkin Numerical Modeling Environment(GNuME) framework, which connects the work in [@NUMA] and the present model. The GNuME framework employs an arbitrary polynomial basis function expansion and provides a choice between continuous Galerkin (CG) and discontinuous Galerkin (DG) methods [@abdi2016efficient]. The framework was previously used to construct the Non-Hydrostatic Unified Model of the Ocean [@kopera2023non]. h-NUMO is implemented in Fortran 90 and uses MPI for parallel computing, making it ideal for high-performance computing environments.
+hNUMO is a high-order isopycnal ocean circulation model that uses barotropic–baroclinic splitting to separate fast and slow motions into coupled subsystems. Extending the work of [@higdon2015multiple] to two horizontal dimensions, hNUMO applies this methodology to two horizontal dimensions using an arbitrary-order, nodal discontinuous Galerkin (DG) discretization. 
 
-h-NUMO is an isopycnal model, meaning it approximates the ocean as a stack of layers, each with a constant density specified by the user. Other widely used vertical coordinates in models of the general ocean circulation include level ($z$) and terrain-fitted ($\sigma$) [@griffies2000developments,@chassignet2006generalized]. Some models, including Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)), use a hybrid coordinate system that employs all of these, with different coordinates in different regions.
+As an isopycnal model, hNUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments,@chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
+
+The model's performance is assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYbrid Coordinate Ocean Model (HYCOM). While hNUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, hNUMO exhibits strong parallel scalability and sustains high efficiency even when operated with a limited number of grid elements per computational core. A detailed description of the model and its numerical discretization as well as model performance is provided in [@GAHOUNZO2026114496].
 
 # Statement of need
 
-Most ocean models employ finite difference or finite volume methods, but these approaches have limitations in terms of accuracy and flexibility. Considering the various timescales of relevant ocean processes, ocean modeling is intrinsically multiscale, and accurately representing these physical processes presents computational challenges. The numerical methods used in ocean models should include desirable features such as low artificial dissipation, efficient resolution of localized flow features, and the capability to handle complex coastline geometries.
[…]
