+++ paper/paper.bib
@@ -13,15 +13,15 @@
 }
 
 @article{abdi2016efficient,
-  title = {Efficient construction of unified continuous and discontinuous Galerkin formulations for the 3D Euler equations},
-  journal = {Journal of Computational Physics},
-  volume = {320},
-  pages = {46-68},
-  year = {2016},
-  issn = {0021-9991},
-  doi = {https://doi.org/10.1016/j.jcp.2016.05.033},
-  url = {https://www.sciencedirect.com/science/article/pii/S0021999116301784},
-  author = {Daniel S. Abdi and Francis X. Giraldo},
+  title    = {Efficient construction of unified continuous and discontinuous Galerkin formulations for the 3D Euler equations},
+  journal  = {Journal of Computational Physics},
+  volume   = {320},
+  pages    = {46-68},
+  year     = {2016},
+  issn     = {0021-9991},
+  doi      = {https://doi.org/10.1016/j.jcp.2016.05.033},
+  url      = {https://www.sciencedirect.com/science/article/pii/S0021999116301784},
+  author   = {Daniel S. Abdi and Francis X. Giraldo},
   keywords = {NUMA, Continuous Galerkin, Discontinuous Galerkin},
   abstract = {A unified approach for the numerical solution of the 3D hyperbolic Euler equations using high order methods, namely continuous Galerkin (CG) and discontinuous Galerkin (DG) methods, is presented. First, we examine how classical CG that uses a global storage scheme can be constructed within the DG framework using constraint imposition techniques commonly used in the finite element literature. Then, we implement and test a simplified version in the Non-hydrostatic Unified Model of the Atmosphere (NUMA) for the case of explicit time integration and a diagonal mass matrix. Constructing CG within the DG framework allows CG to benefit from the desirable properties of DG such as, easier hp-refinement, better stability etc. Moreover, this representation allows for regional mixing of CG and DG depending on the flow regime in an area. The different flavors of CG and DG in the unified implementation are then tested for accuracy and performance using a suite of benchmark problems representative of cloud-resolving scale, meso-scale and global-scale atmospheric dynamics. The value of our unified approach is that we are able to show how to carry both CG and DG methods within the same code and also offer a simple recipe for modifying an existing CG code to DG and vice versa.}
 }
@@ -30,99 +30,133 @@
   title     = {Non-hydrostatic unified model of the ocean with application to ice/ocean interaction modeling},
   author    = {Kopera, Michal A and Gahounzo, Yao and Enderlin, Ellyn M and Giraldo, Francis X and Maslowski, Wieslaw},
   journal   = {GEM-International Journal on Geomathematics},
-  url = {https://doi.org/10.1007/s13137-022-00212-7},
-  doi = {10.1007/s13137-022-00212-7},
+  url       = {https://doi.org/10.1007/s13137-022-00212-7},
+  doi       = {10.1007/s13137-022-00212-7},
   volume    = {14},
   number    = {1},
   pages     = {2},
   year      = {2023},
   publisher = {Springer},
[…]
