### 2026-01-06 b5f1e89
Update on btp/bcl solvers
