+++ src/create_rhs_dynamics_flux.F90
@@ -127,32 +127,24 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
  
    !local variables
  
-   real :: wq
-   integer :: iface, k
- 
    integer :: jj, imm, inbh, ib, kk, ivar, ii
    integer :: imulti
    integer :: multirate
-   real, dimension(4,nq) :: qbl, qbr
-   real, dimension(nq) :: pbl, pbr, H_face_tmp
-   real :: nxl, nyl, nxr, nyr
-   real :: pbpert_edge, c_minus, c_plus
-   real :: c_pb_L, c_pb_R, c_pbub_LR, c_pbub_L, c_pbub_R
-   real, dimension(nq) :: one_plus_eta_edge, c_pb_LR
-   real, dimension(nq) :: flux_edge_x, flux_edge_y
-   integer :: el, iquad, n, I, il, jl, kl, er, itype
-   real, dimension(nq) :: quu, quv, qvu, qvv, ul, ur, vl, vr
-   real :: pU_L, pU_R, lamb, dispu, dispv, flux_x, flux_y, flux
-   real :: hi, H_kx, H_ky
+   integer :: iface, iquad, el, er, il, jl, ir, jr, I, kl, kr, n
+   real :: wq, hi, nxl, nyl, nxr, nyr, un
+   real :: ul, ur, vl, vr, pbl, pbr, clam, one_eta
+   real :: pU_L, pU_R, pbpert_edge
+   real :: qbl(4), qbr(4), flux(3,nq)
+   real :: c_minus, c_plus
+   real :: ppl, ppr, upl, upr, vpl, vpr
    real, dimension(nlayers+1) :: pprime_l, pprime_r
-   real, dimension(nq) :: Q_uu_q, Q_uv_q, Q_vu_q, Q_vv_q, H_bcl_q
-   real :: ppl, ppr, upl, upr, vpl, vpr, pb_cutoff
+   integer :: itype, k
+   real :: H_bcl_ql, H_bcl_qr, H_bcl_q, flux_pb, flux_u, flux_v, fxl, fxr
+   real, dimension(2) :: Qu_ql, Qu_qr, Qv_ql, Qv_qr
 
    jj=1
    kk=1
    imm = 0
-
-   pb_cutoff = (gravity/alpha_mlswe(nlayers))*dry_cutoff
 
    do inbh = 1, num_nbh
       do ib=1,num_send_recv(inbh)
@@ -168,7 +160,6 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
 
          qbl = 0.0 ; qbr = 0.0
          pbl = 0.0 ; pbr = 0.0
-         Q_uu_q = 0.0; Q_uv_q = 0.0; Q_vu_q = 0.0; Q_vv_q = 0.0 ; H_bcl_q = 0.0
 
          do iquad = 1, nq
 
@@ -180,22 +171,34 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
             do n = 1,ngl
                hi = psiq(n,iquad)
                !Left Element
-               qbl(1:4,iquad) = qbl(1:4,iquad) + hi*q_send(1:4,n,kk)
-               pbl(iquad) = pbl(iquad) + hi*q_send(5,n,kk)
+               qbl(1:4) = qbl(1:4) + hi*q_send(1:4,n,kk)
+               pbl = pbl + hi*q_send(5,n,kk)
                !Right Element
-               qbr(1:4,iquad) = qbr(1:4,iquad) + hi*q_recv(1:4,n,kk)
[…]
