+++ src/create_rhs_dynamics_flux.F90
@@ -139,7 +139,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
    real :: ppl, ppr, upl, upr, vpl, vpr
    real, dimension(nlayers+1) :: pprime_l, pprime_r
    integer :: itype, k
-   real :: H_bcl_ql, H_bcl_qr, H_bcl_q, flux_pb, flux_u, flux_v, fxl, fxr
+   real :: H_bcl_ql, H_bcl_qr, H_bcl_q, flux_pb, flux_u, flux_v, fxl, fxr, flux_edge_x, flux_edge_y
    real, dimension(2) :: Qu_ql, Qu_qr, Qv_ql, Qv_qr
 
    jj=1
@@ -158,9 +158,6 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
          !Store Left Side Variables
          !-------------------------------------
 
-         qbl = 0.0 ; qbr = 0.0
-         pbl = 0.0 ; pbr = 0.0
-
          do iquad = 1, nq
 
             nxl = normal_vector_q(1,iquad,1,iface)
@@ -168,6 +165,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
 
             nxr = -nxl; nyr = -nyl
 
+            qbl = 0.0; qbr = 0.0 ; pbl = 0.0; pbr = 0.0
             do n = 1,ngl
                hi = psiq(n,iquad)
                !Left Element
@@ -178,15 +176,6 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                pbr = pbr + hi*q_recv(5,n,kk)
             end do
 
-            ul = qbl(3)/qbl(1); ur = qbr(3)/qbr(1);
-            vl = qbl(4)/qbl(1); vr = qbr(4)/qbr(1);
-
-            Qu_ql(1) = ul*qbl(3) ; Qu_ql(2) = vl*qbl(3)
-            Qu_qr(1) = ur*qbr(3) ; Qu_qr(2) = vr*qbr(3)
-
-            Qv_ql(1) = ul*qbl(4) ; Qv_ql(2) = vl*qbl(4)
-            Qv_qr(1) = ur*qbr(4) ; Qv_qr(2) = vr*qbr(4)
-
             pU_L = nxl * qbl(3) + nyl * qbl(4)
             pU_R = nxr * qbr(3) + nyr * qbr(4)
 
@@ -194,11 +183,20 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
             c_plus  = sqrt(alpha_mlswe(nlayers) * pbl)
             clam = max(c_minus, c_plus)
 
-            pbpert_edge = 0.5 * (qbl(2) + qbr(2)) + (0.5/clam) * (pU_L + pU_R)
+            pbpert_edge = 0.5*(qbl(2) + qbr(2)) + (0.5/clam) * (pU_L + pU_R)
             one_eta = 1.0 + (pbpert_edge/pbl)
 
-            pprime_l(1) = 0.0; pprime_r(1) = 0.0
-            H_bcl_ql = 0.0; H_bcl_qr = 0.0
+            ul = qbl(3)/qbl(1); ur = qbr(3)/qbr(1)
+            vl = qbl(4)/qbl(1); vr = qbr(4)/qbr(1)
+
+            Qu_ql(1) = ul*qbl(3) ; Qu_ql(2) = vl*qbl(3)
[…]
