+++ src/mod_rhs_btp.F90
@@ -44,7 +44,11 @@ contains
 
         ! Compute RHS viscosity terms
 
-        call btp_create_laplacian(rhs_visc_btp,qb_df)
+        if(method_visc == 1) then
+            call btp_create_laplacian_v2(rhs_visc_btp, qprime_df, qb_df)
+        elseif (method_visc > 1) then
+            call btp_create_laplacian(rhs_visc_btp,qb_df)
+        endif 
 
         rhs(1,:) = massinv(:)*rhs(1,:)
         rhs(2,:) = massinv(:)*(rhs(2,:) + rhs_visc_btp(1,:))
@@ -87,9 +91,7 @@ contains
 
         do Iq = 1, npoin_q
 
-            dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0
-            pp = 0.0; up = 0.0; vp = 0.0; pbq = 0.0
-            pprime(:) = 0.0
+            dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0; pbq = 0.0
 
             do ip = 1,npts
                 I = indexq(ip,Iq)
@@ -102,6 +104,9 @@ contains
                 pbq = pbq + hi*pbprime_df(I)
             end do
 
+            Hq = 0.0 ; pprime(:) = 0.0
+            pp = 0.0; up = 0.0; vp = 0.0
+            
             do k = 1, nlayers
                 do ip = 1,npts
                     I = indexq(ip,Iq)
@@ -175,8 +180,6 @@ contains
             end do
         end do
 
-        ! print *, rhs
-
     end subroutine create_rhs_btp_volume_qdf
 
     subroutine create_btp_fluxes_qdf(rhs,qb,qprime_df)
@@ -198,16 +201,20 @@ contains
         real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
 
         integer :: iface, iquad, el, er, il, jl, ir, jr, I, kl, kr, n
-        real :: wq, hi, nxl, nyl, nxr, nyr, un
-        real :: ul, ur, vl, vr, pbl, pbr, clam, one_eta
-        real :: pU_L, pU_R, pbpert_edge
-        real :: qbl(4), qbr(4), flux(3,nq)
-        real :: c_minus, c_plus
-        real :: ppl, ppr, upl, upr, vpl, vpr
+        real :: wq, hi, nxl, nyl, H_kx, H_ky, flux_u, flux_v, flux_pb, nxr, nyr, un
+        real, dimension(nq) :: quu, quv, qvu, qvv, H_face_tmp
+        real :: pbl, pbr, lambq, one_eta
+        real :: pU_L, pU_R, lamb, dispu, dispv, pbpert_edge
+        real :: qbl(4), qbr(4), lam1, lam2, dispp
+        real :: c_minus, c_plus, c_pb_L, c_pb_R, c_pbub_LR, c_pbub_L, c_pbub_R
+        real, dimension(nq) :: c_pb_LR
+        real :: Q_uu_q, Q_uv_q, Q_vu_q, Q_vv_q
+        real :: ppl, ppr, upl, upr, vpl, vpr, pb_cutoff
         real, dimension(nlayers+1) :: pprime_l, pprime_r
         integer :: itype, k
-        real :: H_bcl_ql, H_bcl_qr, H_bcl_q, flux_pb, flux_u, flux_v, fxl, fxr
+        real :: pkl, pkr, ukl, ukr, vkl, vkr, H_bcl_ql, H_bcl_qr, ul, ur, vl, vr, H_bcl_q, clam
         real, dimension(2) :: Qu_ql, Qu_qr, Qv_ql, Qv_qr
+        real :: flux(3,nq), fxl, fxr, flux_edge_x, flux_edge_y
 
[…]
