+++ src/mod_rhs_btp.F90
@@ -73,9 +73,9 @@ contains
         real, dimension(3,npoin,nlayers), intent(in) :: qprime_df
         real, dimension(3,npoin), intent(out) :: rhs
 
-        real :: sc_x, sc_y, Hq, qu, quv, qvu, qv, H_b, pb_cutoff, pk, uk, vk
-        real :: wq, hi, dhdx, dhdy, coef_fric, tb_u, tb_v, ope, &
-                dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd, pbq, H_bclq
+        real :: sc_x, sc_y, Hq, Qu(2), Qv(2)
+        real :: wq, hi, dhdx, dhdy, tb_u, tb_v, ope, &
+                dp, dpp, udp, vdp, ub, vb, Pstress, Pbstress, ubot, vbot, spd, pbq
         real, dimension(nlayers) :: pp, up, vp
         real, dimension(nlayers+1) :: pprime
         integer :: I, Iq, ip, k
@@ -89,13 +89,9 @@ contains
         Pbstress = (gravity/alpha_mlswe(nlayers)) * 10.0 ! pressure corresponding to 10m depth
                                                          ! at which bottom stress is reduced to 0
 
-        pb_cutoff = (gravity/alpha_mlswe(nlayers))*dry_cutoff
-
         do Iq = 1, npoin_q
 
-            dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0
-            pp = 0.0; up = 0.0; vp = 0.0; pbq = 0.0
-            pprime(:) = 0.0 ; H_b = 0.0
+            dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0; pbq = 0.0
 
             do ip = 1,npts
                 I = indexq(ip,Iq)
@@ -108,11 +104,9 @@ contains
                 pbq = pbq + hi*pbprime_df(I)
             end do
 
-            ! if (dp <= pb_cutoff) then
-            !     dp = pb_cutoff
-            !     udp = 0.0 ; vdp = 0.0 ; dpp = 0.0
-            ! end if
-
+            Hq = 0.0 ; pprime(:) = 0.0
+            pp = 0.0; up = 0.0; vp = 0.0
+            
             do k = 1, nlayers
                 do ip = 1,npts
                     I = indexq(ip,Iq)
@@ -122,16 +116,8 @@ contains
                     vp(k) = vp(k) + hi*qprime_df(3,I,k)
                 enddo
 
-                if (pp(k) <= (gravity/alpha_mlswe(k))*dry_cutoff) then
-                    pp(k) = (gravity/alpha_mlswe(k))*dry_cutoff
-                    up(k) = 0.0 ; vp(k) = 0.0
-                end if
-
                 pprime(k+1) = pprime(k) + pp(k)
-                H_bclq = 0.5*alpha_mlswe(k)*(pprime(k+1)**2 - pprime(k)**2)
-                if (abs(pprime(k+1) - pprime(k)) <= (gravity/alpha_mlswe(k))*dry_cutoff) H_bclq = 0.0
-                
-                H_b = H_b + H_bclq
+                Hq = Hq + 0.5*alpha_mlswe(k)*(pprime(k+1)**2 - pprime(k)**2)
 
             enddo
 
@@ -158,15 +144,15 @@ contains
                         gravity*dp*grad_zbot_quad(2,Iq)
 
             ope = 1.0 + (dpp / pbq)
[…]

[…]
