### 2026-01-07 13bc7df
Fix bug in communication routines
