+++ src/create_rhs_dynamics_flux.F90
@@ -468,7 +468,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
    integer :: isub, ic, jj, im, imm, inbh, ib, kk, ivar, imulti
    integer :: multirate
    
-   real, dimension(nlayers) :: alpha_over_g, g_over_alpha, dpp
+   real, dimension(nlayers) :: alpha_over_g, g_over_alpha
    real, dimension(2,nlayers+1) :: p_face, z_face
    real, dimension(nlayers+1) :: p_edge_plus, p_edge_minus, p2l, p2r, z_edge_plus, z_edge_minus
    real, dimension(3,nlayers) :: ql, qr
@@ -484,8 +484,8 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
    real, parameter :: eps1 = 1.0e-20 !  Parameter used to prevent division by zero.
    integer :: il, jl, ir, jr, kl, kr, jquad, n, m, index
    real :: wq, hi, hlx_k, hly_k, hrx_k, hry_k, flux_x, flux_y, hx_k, hy_k, flux
-   real, dimension(2,nq,nlayers) :: H_face, udp_flux, vdp_flux, dp_flux
-   real :: dp_lr(2,nlayers), dp_deficit(2)
+   real, dimension(2,nq,nlayers) :: udp_flux, vdp_flux, dp_flux
+   real :: dp_lr(2,nlayers), dp_deficit(2), H_face(nq,nlayers)
 
    jj=1
    kk=1
@@ -540,7 +540,6 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                vl = ql(3,k) + qbl(3,iquad)
                vr = qr(3,k) + qbr(3,iquad)
 
-               dpp(k) = 0.5*(dpl + dpr)
                uu = 0.5*(ul+ur)
                vv = 0.5*(vl+vr)
                udpl(k) = ul*dpl
@@ -567,14 +566,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                   vdp_flux(2,iquad,k) = vv * (vr*dpr)
                endif
 
-               flux_dp(iquad,k) = nxl*dp_flux(1,iquad,k) + nyl*dp_flux(2,iquad,k)
-
             enddo !k
-
-            ! sum_layer_mass_flux_face(1,iquad,iface) = sum_layer_mass_flux_face(1,iquad,iface)  + &
-            !                                           sum(udp_flux(1,iquad,:))
-            ! sum_layer_mass_flux_face(2,iquad,iface) = sum_layer_mass_flux_face(2,iquad,iface)  + &
-            !                                           sum(vdp_flux(1,iquad,:))
 
             uu_dp_flux_deficit(1) = Qu_face_ave(1,iquad,iface) - sum(udp_flux(1,iquad,:))
             uu_dp_flux_deficit(2) = Qu_face_ave(2,iquad,iface) - sum(udp_flux(2,iquad,:))
@@ -714,10 +706,8 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                nxl = normal_vector_q(1,iquad,1,iface)
                nyl = normal_vector_q(2,iquad,1,iface)
 
-               hlx_k = nxl*H_face(1,iquad,k)
-               hrx_k = nxl*H_face(2,iquad,k)
-               hly_k = nyl*H_face(1,iquad,k)
-               hry_k = nyl*H_face(2,iquad,k)
+               hlx_k = nxl*H_face(iquad,k)
[…]
