+++ src/mod_create_rhs_mlswe.F90
@@ -21,7 +21,7 @@ module mod_create_rhs_mlswe
 
     public :: layer_momentum_rhs, &
               interpolate_layer_from_quad_to_node, rhs_layer_shear_stress, layer_mass_rhs, &
-              consistency_mass_rhs, bcl_rhs
+              bcl_rhs
               
     contains
 
@@ -111,32 +111,6 @@ module mod_create_rhs_mlswe
         end do
 
     end subroutine bcl_rhs
-
-    subroutine consistency_mass_rhs(dp_advec, dprime_df)
-
-        use mod_metrics, only: massinv
-        use mod_input, only: nlayers
-        use mod_grid, only: npoin, npoin_q, nface
-        use mod_basis, only: nq
-
-        implicit none
-
-        real, dimension(npoin, nlayers), intent(out) :: dp_advec
-        real, dimension(npoin, nlayers), intent(in) :: dprime_df
-        
-        integer :: k
-
-        call create_consistency_precommunicator(dprime_df)
-
-        ! Compute the mass advection term for the degree of freedom for dp in each layer
-        call create_consistency_volume_mass(dp_advec, dprime_df)
-
-        ! Compute the mass flux term 
-        call create_consistency_mass_flux(dp_advec, dprime_df)
-
-        call create_postcommunicator_consistency(dp_advec)
-
-    end subroutine consistency_mass_rhs
 
     subroutine interpolate_layer_from_quad_to_node(q_df,q)
 
@@ -339,7 +313,7 @@ module mod_create_rhs_mlswe
         real :: temp_dp, temp_u, temp_v, Ptop_k, Pbot_k, tempbot, Pbstress
         real :: weight, acceleration, pbq
         real, dimension(3) :: qp, qb
-        real, dimension(nlayers) :: temp_uu, temp_vv, H_tmp, u_udp, v_vdp, udp, vdp
+        real, dimension(nlayers) :: temp_uu, temp_vv, H_tmp, u_udp, v_vdp
         real, dimension(2,nlayers) :: u_vdp
         real :: p_tmp(nlayers+1), u, v, dp, weightq, one_over_sumuq, one_over_sumvq
         real :: uu_dp_deficitq, uv_dp_deficitq, vv_dp_deficitq, gradz(2,nlayers+1)
@@ -386,10 +360,10 @@ module mod_create_rhs_mlswe
                 u = qp(2) + qb(2)
                 v = qp(3) + qb(3)
 
-                u_udp(k) = u*udp(k)
-                v_vdp(k) = v*vdp(k)
-                u_vdp(1,k) = v * udp(k)
-                u_vdp(2,k) = u * vdp(k)
+                u_udp(k) = dp * u*u
+                v_vdp(k) = dp * v*v
+                u_vdp(1,k) = u * v * dp
+                u_vdp(2,k) = v * u * dp
 
                 temp_uu(k) = abs(dp*u) + eps1
                 temp_vv(k) = abs(dp*v) + eps1
@@ -523,6 +497,7 @@ module mod_create_rhs_mlswe
         rhs = 0.0
         bot_layer = 0.0
         pprime_temp = 0.0
+        sum_layer_mass_flux = 0.0
 
         Pstress = (gravity/alpha_mlswe(1)) * 50.0 ! pressure corresponding to 50m depth 
                                                   ! at which wind stress is reduced to 0
@@ -587,6 +562,9 @@ module mod_create_rhs_mlswe
             uv_dp_deficitq = Quv_ave(Iq) - sum(u_vdp(1,:))
             vv_dp_deficitq = Qv_ave(Iq) - sum(v_vdp(:))
 
[…]

[…]
