### 2026-01-07 57dcdec
Modified the btp/bcl mass consistency formulation
