+++ src/mod_create_rhs_mlswe.F90
@@ -1168,20 +1168,20 @@ module mod_create_rhs_mlswe
                 vv_dp_flux_deficit(1) = Qv_face_ave(1,iquad,iface) - sum(vdp_flux(1,iquad,:))
                 vv_dp_flux_deficit(2) = Qv_face_ave(2,iquad,iface) - sum(vdp_flux(2,iquad,:))
 
-                dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(dp_flux(1,iquad,:))
-                dp_deficit(2) = btp_mass_flux_face_ave(2,iquad,iface) - sum(dp_flux(2,iquad,:))
+                ! dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(dp_flux(1,iquad,:))
+                ! dp_deficit(2) = btp_mass_flux_face_ave(2,iquad,iface) - sum(dp_flux(2,iquad,:))
 
                 do k = 1,nlayers
 
-                    weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
-                    if (dp_deficit(1)*nxl < 0.0) &
-                        weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
-                    dp_flux(1,iquad,k) = dp_flux(1,iquad,k) + weight * dp_deficit(1)
+                    ! weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
+                    ! if (dp_deficit(1)*nxl < 0.0) &
+                    !     weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
+                    ! dp_flux(1,iquad,k) = dp_flux(1,iquad,k) + weight * dp_deficit(1)
 
-                    weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
-                    if (dp_deficit(2)*nyl < 0.0) &
-                        weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
-                    dp_flux(2,iquad,k) = dp_flux(2,iquad,k) + weight * dp_deficit(2)
+                    ! weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
+                    ! if (dp_deficit(2)*nyl < 0.0) &
+                    !     weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
+                    ! dp_flux(2,iquad,k) = dp_flux(2,iquad,k) + weight * dp_deficit(2)
 
                     ! Adjust the fluxes for the u-momentum equation
                     !x-direction
@@ -1369,6 +1369,16 @@ module mod_create_rhs_mlswe
                     flux_vr(iquad,k) = nxl*vdp_flux(1,iquad,k) + nyl*(vdp_flux(2,iquad,k) + H_face(2,iquad,k))
                 enddo
 
+                dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(flux_dp(iquad,:))
+
+                do k = 1,nlayers
+
+                    weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
+                    if (dp_deficit(1) < 0.0) &
+                        weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
+                    flux_dp(iquad,k) = flux_dp(iquad,k) + weight * dp_deficit(1)
+                enddo
+
             end do ! iquad
             
             ! Do Gauss-Lobatto Integration
@@ -1497,7 +1507,7 @@ module mod_create_rhs_mlswe
[…]
