+++ src/create_rhs_dynamics_flux.F90
@@ -574,20 +574,20 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
             vv_dp_flux_deficit(1) = Qv_face_ave(1,iquad,iface) - sum(vdp_flux(1,iquad,:))
             vv_dp_flux_deficit(2) = Qv_face_ave(2,iquad,iface) - sum(vdp_flux(2,iquad,:))
 
-            dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(dp_flux(1,iquad,:))
-            dp_deficit(2) = btp_mass_flux_face_ave(2,iquad,iface) - sum(dp_flux(2,iquad,:))
+            ! dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(dp_flux(1,iquad,:))
+            ! dp_deficit(2) = btp_mass_flux_face_ave(2,iquad,iface) - sum(dp_flux(2,iquad,:))
 
             do k = 1,nlayers
 
-               weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
-               if (dp_deficit(1)*nxl < 0.0) &
-                  weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
-               dp_flux(1,iquad,k) = dp_flux(1,iquad,k) + weight * dp_deficit(1)
+               ! weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
+               ! if (dp_deficit(1)*nxl < 0.0) &
+               !    weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
+               ! dp_flux(1,iquad,k) = dp_flux(1,iquad,k) + weight * dp_deficit(1)
 
-               weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
-               if (dp_deficit(2)*nyl < 0.0) &
-                  weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
-               dp_flux(2,iquad,k) = dp_flux(2,iquad,k) + weight * dp_deficit(2)
+               ! weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
+               ! if (dp_deficit(2)*nyl < 0.0) &
+               !    weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
+               ! dp_flux(2,iquad,k) = dp_flux(2,iquad,k) + weight * dp_deficit(2)
 
                ! Adjust the fluxes for the u-momentum equation
                ! x-direction
@@ -702,6 +702,16 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                flux_vl(iquad,k) = nxl*vdp_flux(1,iquad,k) + nyl*(vdp_flux(2,iquad,k) + H_face(iquad,k))
             enddo
 
+            dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(flux_dp(iquad,:))
+
+            do k = 1,nlayers
+
+               weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
+               if (dp_deficit(1) < 0.0) &
+                  weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
+               flux_dp(iquad,k) = flux_dp(iquad,k) + weight * dp_deficit(1)
+            enddo
+
          end do ! iquad
          
          ! Do Gauss-Lobatto Integration
[…]
