### 2026-01-08 945542a
Code clean up
