+++ src/create_rhs_dynamics_flux.F90
@@ -239,11 +239,7 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
 
             flux_edge_x = 0.5*(qbl(3) + qbr(3)) + (0.5*clam)*(nxl * qbl(2) + nxr * qbr(2))
             flux_edge_y = 0.5*(qbl(4) + qbr(4)) + (0.5*clam)*(nyl * qbl(2) + nyr * qbr(2))
-            ! flux(1,iquad) = nxl*flux_edge_x(iquad) + nyl*flux_edge_y(iquad)
-
-            fxl = nxl*qbl(3) + nyl*qbl(4)
-            fxr = nxr*qbr(3) + nyr*qbr(4)
-            flux(1,iquad) = 0.5*(fxl - fxr) - 0.5*clam*(qbr(2) - qbl(2))
+            flux(1,iquad) = nxl*flux_edge_x + nyl*flux_edge_y
 
             ! Compute pressure forcing H_face at each element face.
             H_bcl_q = (one_eta**2) * (0.5 * (H_bcl_ql + H_bcl_qr))
@@ -252,9 +248,6 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
 
             btp_mass_flux_face_ave(1,iquad,iface) = btp_mass_flux_face_ave(1,iquad,iface) + flux_edge_x
             btp_mass_flux_face_ave(2,iquad,iface) = btp_mass_flux_face_ave(2,iquad,iface) + flux_edge_y
-
-            ! btp_mass_flux_face_ave(1,iquad,iface) = btp_mass_flux_face_ave(1,iquad,iface) + flux(1,iquad)
-            ! btp_mass_flux_face_ave(2,iquad,iface) = btp_mass_flux_face_ave(2,iquad,iface) + flux(1,iquad)
 
             H_face_ave(iquad,iface) = H_face_ave(iquad,iface) + H_bcl_q
             Qu_face_ave(1,iquad,iface) = Qu_face_ave(1,iquad,iface) + 0.5*(Qu_ql(1) + Qu_qr(1))
@@ -536,6 +529,9 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
                ! Left side of the edge
                dpl = qbl(1,iquad) * ql(1,k)
                dpr = qbr(1,iquad) * qr(1,k)
+               dp_lr(1,k) = dpl
+               dp_lr(2,k) = dpr
+
                ul = ql(2,k) + qbl(2,iquad)
                ur = qr(2,k) + qbr(2,iquad)
                vl = ql(3,k) + qbl(3,iquad)
@@ -574,20 +570,17 @@ subroutine create_nbhs_face_quad(q_face,q_send,q_recv,nvarb,multirate)
             vv_dp_flux_deficit(1) = Qv_face_ave(1,iquad,iface) - sum(vdp_flux(1,iquad,:))
             vv_dp_flux_deficit(2) = Qv_face_ave(2,iquad,iface) - sum(vdp_flux(2,iquad,:))
 
-            ! dp_deficit(1) = btp_mass_flux_face_ave(1,iquad,iface) - sum(dp_flux(1,iquad,:))
-            ! dp_deficit(2) = btp_mass_flux_face_ave(2,iquad,iface) - sum(dp_flux(2,iquad,:))
-
             do k = 1,nlayers
 
-               ! weight = dp_lr(1,k) / (sum(abs(dp_lr(1,:))+eps1))
-               ! if (dp_deficit(1)*nxl < 0.0) &
-               !    weight = dp_lr(2,k) / (sum(abs(dp_lr(2,:))+eps1))
[…]
