### 2026-01-08 f24486f
Concurrence computation in btp solver
