+++ src/mod_rhs_btp.F90
@@ -89,68 +89,167 @@ contains
         Pbstress = (gravity/alpha_mlswe(nlayers)) * 10.0 ! pressure corresponding to 10m depth
                                                          ! at which bottom stress is reduced to 0
 
-        do Iq = 1, npoin_q
+        ! do Iq = 1, npoin_q
 
-            dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0; pbq = 0.0
+        !     dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0; pbq = 0.0
 
-            do ip = 1,npts
-                I = indexq(ip,Iq)
+        !     do ip = 1,npts
+        !         I = indexq(ip,Iq)
+        !         hi = psih(ip,Iq)
+
+        !         dp = dp + hi*qb_df(1,I)
+        !         dpp = dpp + hi*qb_df(2,I)
+        !         udp = udp + hi*qb_df(3,I)
+        !         vdp = vdp + hi*qb_df(4,I)
+        !         pbq = pbq + hi*pbprime_df(I)
+        !     end do
+
+        !     Hq = 0.0 ; pprime(:) = 0.0
+        !     pp = 0.0; up = 0.0; vp = 0.0
+            
+        !     do k = 1, nlayers
+        !         do ip = 1,npts
+        !             I = indexq(ip,Iq)
+        !             hi = psih(ip,Iq)
+        !             pp(k) = pp(k) + hi*qprime_df(1,I,k)
+        !             up(k) = up(k) + hi*qprime_df(2,I,k)
+        !             vp(k) = vp(k) + hi*qprime_df(3,I,k)
+        !         enddo
+
+        !         pprime(k+1) = pprime(k) + pp(k)
+        !         Hq = Hq + 0.5*alpha_mlswe(k)*(pprime(k+1)**2 - pprime(k)**2)
+
+        !     enddo
+
+        !     wq = wjac(Iq)
+        !     ub = udp/dp; vb = vdp/dp
+
+        !     if (botfr == 1) then
+        !         ubot = up(nlayers) + ub
+        !         vbot = vp(nlayers) + vb
+        !         spd = (cd_mlswe/gravity)*pp(nlayers)
+        !         tb_u = spd*ubot
+        !         tb_v = spd*vbot
+        !     elseif (botfr == 2) then
+        !         ubot = up(nlayers) + ub
+        !         vbot = vp(nlayers) + vb
+        !         spd = (cd_mlswe/alpha_mlswe(nlayers))*sqrt(ubot**2 + vbot**2)
+        !         tb_u = spd*ubot
+        !         tb_v = spd*vbot
+        !     end if
+
+        !     sc_x = coriolis_quad(Iq)*vdp + gravity*(tau_wind(1,Iq) - tb_u) - &
+        !                 gravity*dp*grad_zbot_quad(1,Iq)
+        !     sc_y = -coriolis_quad(Iq)*udp + gravity*(tau_wind(2,Iq) - tb_v) - &
+        !                 gravity*dp*grad_zbot_quad(2,Iq)
+
+        !     ope = 1.0 + (dpp / pbq)
+        !     Hq = (ope**2) * Hq
+
+        !     Qu(1) = ub * udp + ope * sum(up(:)*(pp(:) * up(:)))
+        !     Qu(2) = vb * udp + ope * sum(vp(:)*(pp(:) * up(:)))
+        !     Qv(1) = ub * vdp + ope * sum(up(:)*(pp(:) * vp(:)))
+        !     Qv(2) = vb * vdp + ope * sum(vp(:)*(pp(:) * vp(:)))
+
[…]
