### 2026-01-09 0ade43c
Concurrence computation in bcl solver
