+++ src/mod_create_rhs_mlswe.F90
@@ -337,128 +337,112 @@ module mod_create_rhs_mlswe
                                         (sqrt(ope2_ave_df(:))*qprime_df(1,:,k))
         end do
 
-        do Iq = 1, npoin_q
+        do concurrent (Iq = 1:npoin_q)
 
-            p_tmp(1) = 0.0
-            temp_uu = 0.0; temp_vv = 0.0
-            do k = 1,nlayers
-                qp = 0.0; qb = 0.0
-                do ip = 1,npts
-                    I = indexq(ip,Iq)
+            ! init
+            p_tmp(1)      = 0.0
+
+            qb(1) = ope_ave(Iq)
+            qb(2) = uvb_ave(1,Iq)
+            qb(3) = uvb_ave(2,Iq)
+
+            !---------------------------------------
+            ! Build layer quantities
+            !---------------------------------------
+            do k = 1, nlayers
+
+                qp = 0.0
+                do ip = 1, npts
+                    I  = indexq(ip,Iq)
                     hi = psih(ip,Iq)
                     qp(:) = qp(:) + hi*qprime_df(:,I,k)
-                enddo
-
-                qb(1) = ope_ave(Iq)
-                qb(2) = uvb_ave(1,Iq)
-                qb(3) = uvb_ave(2,Iq)
+                end do
 
                 p_tmp(k+1) = p_tmp(k) + sqrt(ope2_ave(Iq)) * qp(1)
-                H_tmp(k) = 0.5*alpha_mlswe(k) * (p_tmp(k+1)**2 - p_tmp(k)**2)
-                
+                H_tmp(k)   = 0.5*alpha_mlswe(k) * (p_tmp(k+1)**2 - p_tmp(k)**2)
+
                 dp = qp(1) * qb(1)
-                u = qp(2) + qb(2)
-                v = qp(3) + qb(3)
+                u  = qp(2) + qb(2)
+                v  = qp(3) + qb(3)
 
-                u_udp(k) = dp * u*u
-                v_vdp(k) = dp * v*v
-                u_vdp(1,k) = u * v * dp
-                u_vdp(2,k) = v * u * dp
+                u_udp(k)    = dp * u*u
+                v_vdp(k)    = dp * v*v
+                u_vdp(1,k)  = u*v*dp
+                u_vdp(2,k)  = v*u*dp
 
-                temp_uu(k) = abs(dp*u) + eps1
-                temp_vv(k) = abs(dp*v) + eps1
+                temp_uu(k)  = abs(dp*u) + eps1
+                temp_vv(k)  = abs(dp*v) + eps1
+
             end do
 
-            gradz = 0.0 ; pbq = 0.0
-            do ip = 1,npts
+            !---------------------------------------
+            ! gradz and pbq
+            !---------------------------------------
+            gradz(:,:) = 0.0
+            pbq        = 0.0
+            do ip = 1, npts
                 I = indexq(ip,Iq)
                 gradz(1,:) = gradz(1,:) + dpsidx(ip,Iq)*z_elv(I,:)
                 gradz(2,:) = gradz(2,:) + dpsidy(ip,Iq)*z_elv(I,:)
-                pbq = pbq + psih(ip,Iq)*pbprime_df(I)
-            enddo
+                pbq        = pbq        + psih(ip,Iq)*pbprime_df(I)
+            end do
 
-            ! Consistency
-            uu_dp_deficitq = Qu_ave(Iq) - sum(u_udp(:))
+            !---------------------------------------
+            ! Consistency deficits
+            !---------------------------------------
[…]
