+++ paper/paper.bib
@@ -13,7 +13,7 @@
 }
 
 @article{abdi2016efficient,
-  title    = {Efficient construction of unified continuous and discontinuous Galerkin formulations for the 3D Euler equations},
+  title    = {Efficient construction of unified continuous and discontinuous {G}alerkin formulations for the 3D {E}uler equations},
   journal  = {Journal of Computational Physics},
   volume   = {320},
   pages    = {46-68},
@@ -41,7 +41,7 @@
 }
 
 @article{escobar2012high,
-  title    = {High-order discontinuous element-based schemes for the inviscid shallow water equations: Spectral multidomain penalty and discontinuous Galerkin methods},
+  title    = {High-order discontinuous element-based schemes for the inviscid shallow water equations: Spectral multidomain penalty and discontinuous {G}alerkin methods},
   journal  = {Applied Mathematics and Computation},
   volume   = {218},
   number   = {9},
@@ -57,7 +57,7 @@
 
 @article{muller2019strong,
   author   = {Andreas Müller and Michal A Kopera and Simone Marras and Lucas C Wilcox and Tobin Isaac and Francis X Giraldo},
-  title    = {Strong scaling for numerical weather prediction at petascale with the atmospheric model NUMA},
+  title    = {Strong scaling for numerical weather prediction at petascale with the atmospheric model {NUMA}},
   journal  = {The International Journal of High Performance Computing Applications},
   volume   = {33},
   number   = {2},
@@ -85,7 +85,7 @@
 
 @article{chassignet2006generalized,
   author  = {Eric P. Chassignet and Harley E. Hurlburt and Ole Martin Smedstad and George R. Halliwell and Alan J. Wallcraft and E. Joseph Metzger and Brian O. Blanton and Carlos Lozano and Desiraju B. Rao and Patrick J. Hogan and Ashwanth Srinivasan},
-  title   = {Generalized Vertical Coordinates for Eddy-Resolving Global and Coastal Ocean Forecasts},
+  title   = {Generalized {V}ertical {C}oordinates for {E}ddy-{R}esolving {G}lobal and {C}oastal {O}cean {F}orecasts},
   journal = {Oceanography},
   year    = {2006},
   month   = {March},
@@ -96,7 +96,7 @@
 
 
 @article{higdon2015multiple,
-  title    = {Multiple time scales and pressure forcing in discontinuous Galerkin approximations to layered ocean models},
+  title    = {Multiple time scales and pressure forcing in discontinuous {G}alerkin approximations to layered ocean models},
   journal  = {Journal of Computational Physics},
   volume   = {295},
   pages    = {230-260},
@@ -113,13 +113,13 @@
   issn      = {00255718, 10886842},
   url       = {http://www.jstor.org/stable/4100149},
   doi       = {https://doi.org/10.1090/S0025-5718-05-01772-2},
[…]
