+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-hNUMO is a high-order discontinuous-Galerkin (DG) isopycnal ocean circulation model that uses barotropic–baroclinic splitting to separate fast and slow motions into coupled subsystems. Extending the work of [@higdon2015multiple] to two horizontal dimensions, hNUMO applies this methodology to two horizontal dimensions using an arbitrary-order, nodal discontinuous Galerkin discretization. 
+hNUMO is a high-order isopycnal ocean circulation model that uses barotropic–baroclinic splitting to separate fast and slow motions into coupled subsystems. Extending the work of [@higdon2015multiple] to two horizontal dimensions, hNUMO applies an arbitrary-order, nodal discontinuous Galerkin discretization. 
 
 As an isopycnal model, hNUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments,@chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
 
@@ -32,13 +32,13 @@ The model's accuracy and performance is assessed through a series of numerical e
 
 Barotropic–baroclinic splitting is commonly used in layered ocean circulation models, which are typically discretized using finite-difference or finite-volume methods. Although these methods are effective, they exhibit limitations in both accuracy and geometric flexibility. Specifically, implementing high-order finite-volume schemes on unstructured meshes is challenging because they require wide reconstruction stencils.
 
-High-order element-based discontinuous Galerkin (DG) methods represent a promising alternative, as they deliver high-order accuracy, reduced numerical dissipation, and efficient resolution of localized flow features [@escobar2012high]. Nevertheless, the application of DG methods to barotropic–baroclinic splitting in ocean circulation models has been limited. The hNUMO implements an arbitrary-order nodal discontinuous Galerkin discretization for both split subsystems. This approach provides a flexible and accurate framework for ocean circulation modeling, facilitating high-order simulations using non-uniform grids while maintaining the efficiency of barotropic–baroclinic coupling.
+High-order element-based discontinuous Galerkin methods represent a promising alternative, as they deliver high-order accuracy, reduced numerical dissipation, and efficient resolution of localized flow features [@escobar2012high]. Nevertheless, the application of DG methods to barotropic–baroclinic splitting in ocean circulation models has been limited. Therefore, the hNUMO implements an arbitrary-order nodal discontinuous Galerkin discretization for both split subsystems. This approach provides a flexible and accurate framework for ocean circulation modeling, facilitating high-order simulations using non-uniform grids while maintaining the efficiency of barotropic–baroclinic coupling.
 
 # h-NUMO features
 
 hNUMO is implemented in the Galerkin Numerical Modeling Environment (GNUME) framework, which supports arbitrary-order polynomial bases and both continuous and discontinuous Galerkin methods [@abdi2016efficient]. The framework was previously used to develop the Non-Hydrostatic Unified Model of the Ocean [@kopera2023non]. The model is written in Fortran 90 and parallelized with MPI for high-performance computing. It is simple and easy to use for idealized and large-scale ocean applications for high-resolution and accurate results.
 
[…]
