### 2026-01-11 545659b
Update on rk bcl solver
