+++ src/mod_create_rhs_mlswe.F90
@@ -25,7 +25,7 @@ module mod_create_rhs_mlswe
               
     contains
 
-    subroutine layer_momentum_rhs(rhs_mom, rhs_visc, qprime_df, q_df)
+    subroutine layer_momentum_rhs(rhs_mom, qprime_df, q_df)
 
         use mod_input, only: nlayers
         use mod_metrics, only: massinv
@@ -34,7 +34,6 @@ module mod_create_rhs_mlswe
 
         implicit none
         real, dimension(2,npoin,nlayers), intent(out) :: rhs_mom
-        real, dimension(2,npoin,nlayers), intent(in) :: rhs_visc
         real, dimension(3,npoin,nlayers), intent(in) :: q_df, qprime_df
 
         integer :: k,I
@@ -46,11 +45,6 @@ module mod_create_rhs_mlswe
         call Apply_layers_fluxes(rhs_mom, qprime_df)
 
         call bcl_create_postcommunicator_momentum(rhs_mom)
-
-        do k = 1, nlayers
-            rhs_mom(1,:,k) = massinv(:)*rhs_mom(1,:,k) + rhs_visc(1,:,k)
-            rhs_mom(2,:,k) = massinv(:)*rhs_mom(2,:,k) + rhs_visc(2,:,k)
-        end do
 
     end subroutine layer_momentum_rhs
 
@@ -86,6 +80,7 @@ module mod_create_rhs_mlswe
     subroutine bcl_rhs(rhs, qprime_df, q_df)
 
         use mod_input, only: nlayers
+        use mod_metrics, only: massinv
         use mod_grid, only: npoin, npoin_q, intma, intma_dg_quad, nface
         use mod_basis, only: nq, ngl
 
@@ -329,112 +324,127 @@ module mod_create_rhs_mlswe
                                         (sqrt(ope2_ave_df(:))*qprime_df(1,:,k))
         end do
 
-        do concurrent (Iq = 1:npoin_q)
+        do concurrent(Iq = 1:npoin_q)
 
-            ! init
-            p_tmp(1)      = 0.0
+            p_tmp(:) = 0.0
+            pprime_temp(:) = 0.0
 
-            qb(1) = ope_ave(Iq)
-            qb(2) = uvb_ave(1,Iq)
-            qb(3) = uvb_ave(2,Iq)
-
-            !---------------------------------------
-            ! Build layer quantities
-            !---------------------------------------
-            do k = 1, nlayers
-
-                qp = 0.0
-                do ip = 1, npts
-                    I  = indexq(ip,Iq)
+            temp_uu = 0.0; temp_vv = 0.0
+            do k = 1,nlayers
+                qp = 0.0; qb = 0.0
+                do ip = 1,npts
+                    I = indexq(ip,Iq)
                     hi = psih(ip,Iq)
                     qp(:) = qp(:) + hi*qprime_df(:,I,k)
-                end do
+                enddo
+
+                qb(1) = ope_ave(Iq)
+                qb(2) = uvb_ave(1,Iq)
+                qb(3) = uvb_ave(2,Iq)
 
                 p_tmp(k+1) = p_tmp(k) + sqrt(ope2_ave(Iq)) * qp(1)
-                H_tmp(k)   = 0.5*alpha_mlswe(k) * (p_tmp(k+1)**2 - p_tmp(k)**2)
-
+                H_tmp(k) = 0.5*alpha_mlswe(k) * (p_tmp(k+1)**2 - p_tmp(k)**2)
+                
                 dp = qp(1) * qb(1)
-                u  = qp(2) + qb(2)
-                v  = qp(3) + qb(3)
+                u = qp(2) + qb(2)
+                v = qp(3) + qb(3)
 
-                u_udp(k)    = dp * u*u
-                v_vdp(k)    = dp * v*v
[…]
