### 2026-01-13 a9ea957
Update bcl fluxes
