+++ src/mod_gpu.F90
@@ -1,0 +1,115 @@
+module mod_gpu
+
+  use openacc
+  use mpi_utilities
+
+  implicit none
+
+  integer :: numgpu, igpu
+  logical :: is_openacc_initialized=.false.
+
+  public :: numgpu, igpu
+  public :: openacc_initialize, openacc_enter_data
+
+contains
+
+  subroutine openacc_initialize()
+
+#ifdef _OPENACC
+    !Get number of GPUs
+    numgpu = acc_get_num_devices(acc_device_nvidia)
+    if(numgpu <= 0) then
+      print *,'Error in Initialize_OpenACC_Init: numgpu = ',numgpu
+      call exit(-99)
+    end if
+    is_openacc_initialized = .true.
+
+    !Set the GPU device mapped to this MPI rank
+    igpu = mod(irank, numgpu)
+    call acc_set_device_num(igpu, acc_device_nvidia)
+
+    print*,'------------------------------------------'
+    write(*,'("irank igpu nproc numgpu = ",4(i7,1x))') irank,igpu,nproc,numgpu
+    print*,'------------------------------------------'
+    if(igpu < 0) then
+      print *,'Error in Initialize_OpenACC_Init: igpu = ',igpu
+      call exit(-99)
+    end if
+#endif
+
+  end subroutine openacc_initialize
+
+  subroutine openacc_enter_data(sim)
+    use mod_type_numa_simulation
+    type(numa_simulation), intent(inout) :: sim(nsim_rank)
+
+    !$acc enter data copyin(sim(1), sim(1)%q0, sim(1)%q, sim(1)%press, sim(1)%sd )
+    
+    !$acc enter data copyin( sim(1)%ti, &
+    !$acc           sim(1)%ti%itime, sim(1)%ti%time, sim(1)%ti%Nstages, sim(1)%ti%dt, &
+    !$acc           sim(1)%ti%delta, sim(1)%ti%lambda, &
+    !$acc           sim(1)%ti%q1, sim(1)%ti%rhs, sim(1)%ti%RK_a, sim(1)%ti%RK_beta, &
+    !$acc           sim(1)%ti%A, sim(1)%ti%At, sim(1)%ti%b, sim(1)%ti%q_tt, sim(1)%ti%b_tt, &
+    !$acc           sim(1)%ti%q_ARK_hat, sim(1)%ti%q_ARK_hat_correction, sim(1)%ti%q_Stage, sim(1)%ti%q_Stage_R)
+
+    !$acc enter data copyin(sim(1)%ti%sol, &
+    !$acc           sim(1)%ti%sol%nkrylov, sim(1)%ti%sol%maxiter, sim(1)%ti%sol%etol, &
+    !$acc           sim(1)%ti%sol%krylov, sim(1)%ti%sol%krylow, sim(1)%ti%sol%weight, &
+    !$acc           sim(1)%ti%sol%gamma0, sim(1)%ti%sol%gamma0_norm, sim(1)%ti%sol%gamma1, &
+    !$acc           sim(1)%ti%sol%solver_tol, sim(1)%ti%sol%ndof, &
+    !$acc           sim(1)%ti%sol%tvec1, sim(1)%ti%sol%tvec2, sim(1)%ti%sol%tscal, sim(1)%ti%sol%tscal_sum, &
[…]
