+++ src/mod_create_rhs_mlswe.F90
@@ -318,7 +318,7 @@ module mod_create_rhs_mlswe
         real :: p_tmp(nlayers+1), u, v, dp, weightq, one_over_sumuq, one_over_sumvq
         real :: uu_dp_deficitq, uv_dp_deficitq, vv_dp_deficitq, gradz(2,nlayers+1)
         real :: z_elv(npoin,nlayers+1)
-        real, parameter :: eps1 = 1.0e-20!  Parameter used to prevent division by zero.
+        real, parameter :: eps1 = 1.0e-10 !  Parameter used to prevent division by zero.
         real :: flux(2,2)
 
         rhs_mom = 0.0
@@ -383,23 +383,23 @@ module mod_create_rhs_mlswe
             do ip = 1, npts
                 I = indexq(ip,Iq)
                 ! z(:) = 0.0
-                ! z(nlayers+1) = zbot_df(I)
-                ! do k = nlayers,1,-1
-                !     z(k) = z(k+1) + (alpha_mlswe(k)/gravity)*(sqrt(ope2_ave_df(I))*qprime_df(1,I,k))
+                z(nlayers+1) = zbot_df(I)
+                do k = nlayers,1,-1
+                    z(k) = z(k+1) + (alpha_mlswe(k)/gravity)*(sqrt(ope2_ave_df(I))*qprime_df(1,I,k))
 
-                !     gradz(1,k) = gradz(1,k) + dpsidx(ip,Iq) * z(k)
-                !     gradz(2,k) = gradz(2,k) + dpsidy(ip,Iq) * z(k)
-                ! enddo 
+                    gradz(1,k) = gradz(1,k) + dpsidx(ip,Iq) * z(k)
+                    gradz(2,k) = gradz(2,k) + dpsidy(ip,Iq) * z(k)
+                enddo 
 
                 ! do k = 1, nlayers
                 !     gradz(1,k) = gradz(1,k) + dpsidx(ip,Iq) * z_tmp(k)
                 !     gradz(2,k) = gradz(2,k) + dpsidy(ip,Iq) * z_tmp(k)
                 ! end do
 
-                do k = 1, nlayers
-                    gradz(1,k) = gradz(1,k) + dpsidx(ip,Iq) * z_elv(I,k)
-                    gradz(2,k) = gradz(2,k) + dpsidy(ip,Iq) * z_elv(I,k)
-                end do
+                ! do k = 1, nlayers
+                !     gradz(1,k) = gradz(1,k) + dpsidx(ip,Iq) * z_elv(I,k)
+                !     gradz(2,k) = gradz(2,k) + dpsidy(ip,Iq) * z_elv(I,k)
+                ! end do
 
                 gradz(1,nlayers+1) = grad_zbot_quad(1,Iq)
                 gradz(2,nlayers+1) = grad_zbot_quad(2,Iq)
@@ -495,7 +495,7 @@ module mod_create_rhs_mlswe
         real :: p_tmp(nlayers+1), u, v,weightq, one_over_sumuq, one_over_sumvq
         real :: uu_dp_deficitq, uv_dp_deficitq, vv_dp_deficitq, gradz(2,nlayers+1)
         real :: z_elv(npoin,nlayers+1)
-        real, parameter :: eps1 = 1.0e-20 !  Parameter used to prevent division by zero.
+        real, parameter :: eps1 = 1.0e-10 !  Parameter used to prevent division by zero.
         real :: flux(3,3)
 
         rhs = 0.0
@@ -509,11 +509,11 @@ module mod_create_rhs_mlswe
                                                          ! at which bottom stress is reduced to 0
 
         ! Find layer interfaces
-        z_elv(:,nlayers+1) = zbot_df(:)
[…]
