### 2026-01-26 #9147 @mancellin
Thanks for the updated paper, it looks good to me.

I only have two nitpicks and a typo:
- the name of the code is sometimes spelled "hNUMO" and sometimes "h-NUMO"
- first paragraph (l. 7): "a **high-order** isopycnal ocean circulation model" Is the model high-order or is it only its discretization which is discussed in the next sentence?
- l. 55 "suport" -> "support"
