### 2026-01-26 e6c6629
Clean up and update on btp solver
