### 2026-02-03 7890b98
Update on ti_rk3_bcl routine
