+++ src/ti_rk_bcl.F90
@@ -39,11 +39,10 @@ subroutine ti_rk_bcl(q_df, qb_df)
     
     ! ==================== Prediction step =================================
 
-    call extract_qprime_df_face(qprime_df,q_df,qb_df)
-
     qbp_df = qb_df
-    dpprime_visc(:,:) = qprime_df(1,:,:)
     q_df2 = q_df
+
+    call extract_qprime_df_face(qprime_df,q_df2,qbp_df)
 
     call btp_bcl_coeffs_qdf(qprime_df)
     call ti_barotropic_ssprk_mlswe(qbp_df, qprime_df)
@@ -54,7 +53,6 @@ subroutine ti_rk_bcl(q_df, qb_df)
     call extract_qprime_df_face(qprime_df2,q_df2,qbp_df)
 
     qprime_df2 = 0.5*(qprime_df2 + qprime_df)
-    dpprime_visc(:,:) = qprime_df2(1,:,:)
 
     call btp_bcl_coeffs_qdf(qprime_df2)
     call ti_barotropic_ssprk_mlswe(qb_df,qprime_df2)
