+++ mod_rhs_btp.F90
@@ -8,12 +8,11 @@
 module mod_rhs_btp
 
     use mod_constants, only: gravity
-    use mod_grid, only : npoin_q, npoin, nelem, intma_dg_quad
+    use mod_grid, only : npoin_q, npoin, nelem
     use mod_basis, only: nglx, ngly, nglz, npts, dpsiqx, dpsiqy, dpsiqz, nqx, nqy, nqz, &
                          psiqx, psiqy, psiqz, nq, ngl
     use mod_grid, only: intma, npoin_q, npoin, nface
     use mod_laplacian_quad, only: btp_create_laplacian
-    use mod_barotropic_terms, only: btp_extract_df
     use mod_input, only: nlayers, method_visc
     use mod_metrics, only: ksiq_x, ksiq_y, ksiq_z, &
                            etaq_x, etaq_y, etaq_z, &
@@ -93,8 +92,6 @@ contains
         !$acc           pp,up,vp,pprime,Qu,Qv,dhdx,dhdy)
         do Iq = 1, npoin_q
 
-            wq = wjac(Iq)
-
             ! Quadrature-projected barotropic quantities
             dp = 0.0; dpp = 0.0; udp = 0.0; vdp = 0.0; pbq = 0.0
 
@@ -109,9 +106,6 @@ contains
                 vdp = vdp + hi * qb_df(4,I)
                 pbq = pbq + hi * pbprime_df(I)
             end do
-
-            ub = udp / dp
-            vb = vdp / dp
 
             ! Layer projections + hydrostatic-like Hq
             Hq = 0.0 ; pprime(:)= 0.0
@@ -132,6 +126,11 @@ contains
                 pprime(k+1) = pprime(k) + pp(k)
                 Hq = Hq + 0.5 * alpha_mlswe(k) * (pprime(k+1)**2 - pprime(k)**2)
             end do
+
+            ! Local quantities
+            wq = wjac(Iq)
+            ub = udp / dp
+            vb = vdp / dp
 
             tb_u = 0.0 ; tb_v = 0.0
             if (botfr == 1) then
@@ -223,207 +222,219 @@ contains
         real :: qbl(4), qbr(4), lam1, lam2, dispp
         real :: c_minus, c_plus, c_pb_L, c_pb_R, c_pbub_LR, c_pbub_L, c_pbub_R
         real, dimension(nq) :: c_pb_LR
-        real :: Q_uu_q, Q_uv_q, Q_vu_q, Q_vv_q, pb_cutoff
-        real, dimension(nlayers) :: ppl, ppr, upl, upr, vpl, vpr
+        real :: Q_uu_q, Q_uv_q, Q_vu_q, Q_vv_q
+        real :: ppl, ppr, upl, upr, vpl, vpr, pb_cutoff
         real, dimension(nlayers+1) :: pprime_l, pprime_r
         integer :: itype, k
         real :: pkl, pkr, ukl, ukr, vkl, vkr, H_bcl_ql, H_bcl_qr, ul, ur, vl, vr, H_bcl_q, clam
         real, dimension(2) :: Qu_ql, Qu_qr, Qv_ql, Qv_qr
         real :: flux(3,nq), fxl, fxr, flux_edge_x, flux_edge_y, ope_ppl, ope_ppr
 
-        do iface = 1, nface
+        do concurrent (iface = 1:nface, iquad = 1:nq)
 
             ! Skip boundary faces
             if (face_type(iface) == 2) cycle
 
-            do iquad = 1, nq
+            el = face(7,iface)
+            er = face(8,iface)
 
-                el = face(7,iface)
[…]
