### 2026-02-03 c709c71
Fix typo in the paper
