+++ paper/paper.md
@@ -22,25 +22,25 @@ bibliography: paper.bib
 
 # Summary
 
-hNUMO is a high-order isopycnal ocean circulation model that uses barotropic–baroclinic splitting to separate fast and slow motions into coupled subsystems. Extending the work of [@higdon2015multiple] to two horizontal dimensions, hNUMO applies an arbitrary-order, nodal discontinuous Galerkin discretization. 
+h-NUMO is an isopycnal ocean circulation model that uses barotropic–baroclinic splitting to separate fast and slow motions into coupled subsystems. Extending the work of [@higdon2015multiple] to two horizontal dimensions, h-NUMO applies an arbitrary-order, nodal discontinuous Galerkin discretization. 
 
-As an isopycnal model, hNUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments,@chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
+As an isopycnal model, h-NUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments,@chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
 
-The model's accuracy and performance is assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYCOM model. While hNUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, hNUMO exhibits strong parallel scalability and sustains high efficiency even when run with a fewer number of grid elements per computational core. A detailed description of the model and its numerical discretization as well as model performance is provided in [@GAHOUNZO2026114496].
+The model's accuracy and performance is assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYCOM model. While h-NUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, h-NUMO exhibits strong parallel scalability and sustains high efficiency even when run with a fewer number of grid elements per computational core. A detailed description of the model and its numerical discretization as well as model performance is provided in [@GAHOUNZO2026114496].
 
 # Statement of need
 
 Barotropic–baroclinic splitting is commonly used in layered ocean circulation models, which are typically discretized using finite-difference or finite-volume methods. Although these methods are effective, they exhibit limitations in both accuracy and geometric flexibility. Specifically, implementing high-order finite-volume schemes on unstructured meshes is challenging because they require wide reconstruction stencils.
 
[…]
