+++ paper/paper.bib
@@ -19,7 +19,7 @@
   pages    = {46-68},
   year     = {2016},
   issn     = {0021-9991},
-  doi      = {https://doi.org/10.1016/j.jcp.2016.05.033},
+  doi      = {10.1016/j.jcp.2016.05.033},
   url      = {https://www.sciencedirect.com/science/article/pii/S0021999116301784},
   author   = {Daniel S. Abdi and Francis X. Giraldo},
   keywords = {NUMA, Continuous Galerkin, Discontinuous Galerkin},
@@ -48,7 +48,7 @@
   pages    = {4825-4848},
   year     = {2012},
   issn     = {0096-3003},
-  doi      = {https://doi.org/10.1016/j.amc.2011.10.046},
+  doi      = {10.1016/j.amc.2011.10.046},
   url      = {https://www.sciencedirect.com/science/article/pii/S0096300311012914},
   author   = {J.A. Escobar-Vargas and P.J. Diamessis and F.X. Giraldo},
   keywords = {Shallow water equations, Spectral multidomain methods, Discontinuous Galerkin methods, Interfacial treatment},
@@ -77,7 +77,7 @@
   pages    = {123-192},
   year     = {2000},
   issn     = {1463-5003},
-  doi      = {https://doi.org/10.1016/S1463-5003(00)00014-7},
+  doi      = {10.1016/S1463-5003(00)00014-7},
   url      = {https://www.sciencedirect.com/science/article/pii/S1463500300000147},
   author   = {Stephen M. Griffies and Claus Böning and Frank O. Bryan and Eric P. Chassignet and Rüdiger Gerdes and Hiroyasu Hasumi and Anthony Hirst and Anne-Marie Treguier and David Webb},
   abstract = {This paper presents some research developments in primitive equation ocean models which could impact the ocean component of realistic global coupled climate models aimed at large-scale, low frequency climate simulations and predictions. It is written primarily to an audience of modellers concerned with the ocean component of climate models, although not necessarily experts in the design and implementation of ocean model algorithms.}
@@ -91,7 +91,8 @@
   month   = {March},
   note    = {},
   volume  = {issue_volume},
-  doi     = {https://doi.org/10.5670/oceanog.2006.95}
+  doi     = {10.5670/oceanog.2006.95},
+  url     = {https://doi.org/10.5670/oceanog.2006.95}
 }
 
 
@@ -102,7 +103,7 @@
   pages    = {230-260},
   year     = {2015},
   issn     = {0021-9991},
-  doi      = {https://doi.org/10.1016/j.jcp.2015.04.010},
+  doi      = {10.1016/j.jcp.2015.04.010},
   url      = {https://www.sciencedirect.com/science/article/pii/S0021999115002569},
   author   = {Robert L. Higdon},
   keywords = {Ocean modeling, Multi-layer ocean models, Discontinuous Galerkin, Barotropic–baroclinic splitting, Geostrophic adjustment},
@@ -112,7 +113,7 @@
 @article{ruuth2006global,
   issn      = {00255718, 10886842},
   url       = {http://www.jstor.org/stable/4100149},
-  doi       = {https://doi.org/10.1090/S0025-5718-05-01772-2},
+  doi       = {10.1090/S0025-5718-05-01772-2},
[…]
