### 2026-02-28 #9147 @ygahounzo
@cheginit I have changed the title from h-NUMO: A Hydrostatic Version of the Non-hydrostatic Model of the Ocean to h-NUMO: A high-performance discontinuous-Galerkin code for hydrostatic ocean modelling. I am wondering whether I should use the old title when creating a DOI on Zenodo to match this submission, or if the new title would be fine?
