### 2026-03-01 9bcbc90
Fix typo in the paper
