+++ paper/paper.md
@@ -26,7 +26,7 @@ h-NUMO is an isopycnal ocean circulation model that uses barotropic–baroclinic
 
 As an isopycnal model, h-NUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments,@chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
 
-The model's accuracy and performance is assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYCOM model. While h-NUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, h-NUMO exhibits strong parallel scalability and sustains high efficiency even when run with a fewer number of grid elements per computational core. A detailed description of the model and its numerical discretization as well as model performance is provided in [@GAHOUNZO2026114496].
+The model's accuracy and performance are assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYCOM model. While h-NUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, h-NUMO exhibits strong parallel scalability and sustains high efficiency even when run with a fewer number of grid elements per computational core. A detailed description of the model and its numerical discretization, as well as model performance, is provided in [@GAHOUNZO2026114496].
 
 # Statement of need
 
