### 2026-03-01 #9147 @ygahounzo
@cheginit I have released the new version (version 0.1.1), and the DOI to the Zenodo archive is https://doi.org/10.5281/zenodo.18817049.
