### 2026-03-03 #9147 @cheginit
@ygahounzo I noticed that in the references, there is `issue_volume` instead of the actual volume number on line 70. The correct volume can be found [here](https://tos.org/oceanography/article/generalized-vertical-coordinates-for-eddy-resolving-global-and-coastal-ocea#citation).
