### 2026-03-03 0744c04
Fix volume number in the reference for JOSS paper
