+++ paper/paper.bib
@@ -90,7 +90,9 @@
   year    = {2006},
   month   = {March},
   note    = {},
-  volume  = {issue_volume},
+  volume  = {19},
+  number   = {1},
+  pages    = {118-129},
   doi     = {10.5670/oceanog.2006.95},
   url     = {https://doi.org/10.5670/oceanog.2006.95}
 }
