### 2026-03-03 #9147 @cheginit
@ygahounzo as noted above there are two `higdon2015multiple` entries in your bib file.
