### 2026-03-06 #9147 @kthyng
@ygahounzo
* Can you change the version in your Zenodo metadata to match here (`v0.1.1`)

paper:

1. **Page 1 – Summary**
   Missing parentheses around a citation.
   - Current: `...coordinates Chassignet et al. (2006), hybrid-coordinate models...`
   - Suggested: `...coordinates (Chassignet et al., 2006), hybrid-coordinate models...`

2. **Page 1 – Summary**
   Incorrect phrase “a fewer number of”.
   - Current: `...run with a fewer number of grid elements per computational core.`
   - Suggested: `...run with fewer grid elements per computational core.`

3. **Page 1 – Statement of need**
   Extra article before the model name.
   - Current: `Therefore, the h-NUMO implements...`
   - Suggested: `Therefore, h-NUMO implements...`

4. **Page 2 – h-NUMO features**
   Subject–verb agreement and pluralization.
   - Current: `The model also support external grid mesh...`
   - Suggested: `The model also supports external grid meshes...`

5. **Page 2 – Acknowledgements**
   Spelling error.
   - Current: `horizantal`
   - Suggested: `horizontal`
