### 2026-03-08 49eaab4
Revise paper text and references
