+++ paper/paper.md
@@ -24,15 +24,15 @@ bibliography: paper.bib
 
 h-NUMO is an isopycnal ocean circulation model that uses barotropic–baroclinic splitting to separate fast and slow motions into coupled subsystems. Extending the work of [@higdon2015multiple] to two horizontal dimensions, h-NUMO applies an arbitrary-order, nodal discontinuous Galerkin discretization. 
 
-As an isopycnal model, h-NUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments,@chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
+As an isopycnal model, h-NUMO represents the ocean as a stack of constant-density layers. In contrast to traditional $z$-level and terrain-following ($\sigma$) coordinates [@griffies2000developments; @chassignet2006generalized], hybrid-coordinate models such as Hybrid Coordinate Ocean Model ([HYCOM](https://github.com/HYCOM)) and Modular Ocean Model version 6 ([MOM6](https://github.com/NOAA-GFDL/MOM6)) combine multiple vertical coordinates across different regions.
 
-The model's accuracy and performance are assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYCOM model. While h-NUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, h-NUMO exhibits strong parallel scalability and sustains high efficiency even when run with a fewer number of grid elements per computational core. A detailed description of the model and its numerical discretization, as well as model performance, is provided in [@GAHOUNZO2026114496].
+The model's accuracy and performance are assessed through a series of numerical experiments, including a double-gyre circulation test case, with results benchmarked against those from the HYCOM model. While h-NUMO incurs a higher computational cost per degree of freedom, it attains comparable resolved kinetic energy in substantially less wall-clock time, achieving an order-of-magnitude speedup and requiring fewer computational resources. Additionally, h-NUMO exhibits strong parallel scalability and sustains high efficiency even when run with fewer grid elements per computational core. A detailed description of the model and its numerical discretization, as well as model performance, is provided in [@GAHOUNZO2026114496].
 
 # Statement of need
 
 Barotropic–baroclinic splitting is commonly used in layered ocean circulation models, which are typically discretized using finite-difference or finite-volume methods. Although these methods are effective, they exhibit limitations in both accuracy and geometric flexibility. Specifically, implementing high-order finite-volume schemes on unstructured meshes is challenging because they require wide reconstruction stencils.
 
-High-order element-based discontinuous Galerkin methods represent a promising alternative, as they deliver high-order accuracy, reduced numerical dissipation, and efficient resolution of localized flow features [@escobar2012high]. Nevertheless, the application of DG methods to barotropic–baroclinic splitting in ocean circulation models has been limited. Therefore, the h-NUMO implements an arbitrary-order nodal discontinuous Galerkin discretization for both split subsystems. This approach provides a flexible and accurate framework for ocean circulation modeling, facilitating high-order simulations using non-uniform grids while maintaining the efficiency of barotropic–baroclinic coupling.
[…]
